% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboards.R
\name{add_dashboard_in_group}
\alias{add_dashboard_in_group}
\title{Creates a new empty dashboard on the specified workspace.Required scope: Content.Create}
\usage{
add_dashboard_in_group(token, groupId, name, output = "value")
}
\arguments{
\item{groupId}{OK}
}
\value{
A `data.frame` object.
}
\description{
Creates a new empty dashboard on the specified workspace.Required scope: Content.Create
}
