% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{add_group_user}
\alias{add_group_user}
\title{Grants the specified user permissions to the specified workspace.Notes:}
\usage{
add_group_user(
  token,
  groupId,
  displayName,
  emailAddress,
  groupUserAccessRight,
  identifier,
  principalType,
  output = "value"
)
}
\arguments{
\item{groupId}{OK}
}
\value{
A `data.frame` object.
}
\description{
Grants the specified user permissions to the specified workspace.Notes:
}
