% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{add_power_bi_encryption_key}
\alias{add_power_bi_encryption_key}
\title{Adds an encryption key for Power BI workspaces assigned to a capacity.Note: The user must have administrator rights (such as Office 365 Global Administrator or Power BI Service Administrator) to call this API. This API allows 600 requests per hour at maximum. Required scope: Tenant.Read.All or Tenant.ReadWrite.AllTo set the permissions scope, see Register an app.}
\usage{
add_power_bi_encryption_key(
  token,
  activate,
  isDefault,
  keyVaultKeyIdentifier,
  name
)
}
\arguments{
\item{activate}{Indicates to activate any inactivated capacities to use this key for its encryption}

\item{isDefault}{Indicates that this key is set as default for the entire tenant. Any new capacity creation will inherit this key upon creation}

\item{keyVaultKeyIdentifier}{Uri to the version of the Azure Key Vault key to be used}

\item{name}{The name of the encryption key}
}
\value{
A `data.frame` object.
}
\description{
Adds an encryption key for Power BI workspaces assigned to a capacity.Note: The user must have administrator rights (such as Office 365 Global Administrator or Power BI Service Administrator) to call this API. This API allows 600 requests per hour at maximum. Required scope: Tenant.Read.All or Tenant.ReadWrite.AllTo set the permissions scope, see Register an app.
}
