% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateData.R
\name{annotateData}
\alias{annotateData}
\title{Add a column of annotations to a data frame}
\usage{
annotateData(
  x,
  keyColumnName,
  annotationVector,
  annotationName,
  parseProteinGroups = TRUE,
  parsePTMs = FALSE
)
}
\arguments{
\item{x}{a data frame to which to add annotations}

\item{keyColumnName}{column name in x that matches the names in the annotation vector}

\item{annotationVector}{named vector with names that match to keyColumnName}

\item{annotationName}{name for the new column that will be added to x containing annotations}

\item{parseProteinGroups}{flag indicating if the key column is protein groups that need to be parsed to match lookup vector (default = TRUE)}

\item{parsePTMs}{flag indicating if the key column contains PTM accessions that need to be parsed to match lookup vector (default = FALSE)}
}
\value{
an annotated data frame
}
\description{
Add a column of annotations to a data frame
}
\examples{
results <- read.delim(file = "results.txt", quote = "")
db <- loadUniProtFASTA("SwissProt.fasta")
getProteinNames <- db[["getProteinNames"]]
results_ann <- annotateData(x = results, keyColumnName = "Protein", annotationVector = getProteinNames, annotationName = "Protein.Name")

}
