% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{authenticate}
\alias{authenticate}
\title{Returns authentication token.}
\usage{
authenticate(name, key, secret = NULL)
}
\arguments{
\item{name}{The app name}

\item{key}{The app key}

\item{secret}{The app secret}
}
\value{
A `token` object.
}
\description{
Returns authentication token.
}
