% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{bind_to_gateway}
\alias{bind_to_gateway}
\title{Binds the specified dataset from "My Workspace" to the specified gateway with (optional) given set of datasource Ids. This only supports the On-Premises Data Gateway.Required scope: Dataset.ReadWrite.All}
\usage{
bind_to_gateway(
  token,
  datasetId,
  datasourceObjectIds,
  gatewayObjectId,
  output = "value"
)
}
\arguments{
\item{groupId}{OK}
}
\value{
A `data.frame` object.
}
\description{
Binds the specified dataset from "My Workspace" to the specified gateway with (optional) given set of datasource Ids. This only supports the On-Premises Data Gateway.Required scope: Dataset.ReadWrite.All
}
