% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{cap.barley}
\alias{cap.barley}
\title{Example data from a two-row barley population}
\format{An object of class \code{list} with three elements: 1) \code{pheno} - a
\code{data.frame} with 2104 rows and 4 columns. The first column contains the
line name, the second column contains the trial name, the and the remaining two
columns contain phenotypic data on two quantitative traits; 2) \code{geno} - a
\code{data.frame} with 2187 rows and 1594 columns. The first 4 columns contain
information on the name, allels, chromosome, and genetic
map position (respectively) of 2187 molecular markers. Remaining columns
contain information on the genotypic states at each of those markers for all
of the barley lines; 3) \code{geno_mat} - a
\code{matrix} with 1590 rows and 2187 columns. The elements of the matrix contain
the genotypic states at each of 2187 markers for each of 1590 barley lines.}
\source{
This data was downloaded from the Triticeae Toolbox at https://triticeaetoolbox.org/barley/.
}
\usage{
cap.barley
}
\description{
A \code{list} of phenotypic and molecular marker data on a population of 1590
two-row barley lines.
}
\keyword{datasets}
