% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{capacities_unassignworkspacesfromcapacity}
\alias{capacities_unassignworkspacesfromcapacity}
\title{Unassigns the provided workspaces from capacity.Note: The user must have administrator rights (such as Office 365 Global Administrator or Power BI Service Administrator) to call this API. Required scope: Tenant.Read.All or Tenant.ReadWrite.AllTo set the permissions scope, see Register an app.}
\usage{
capacities_unassignworkspacesfromcapacity(token, workspacesToUnassign)
}
\arguments{
\item{workspacesToUnassign}{Workspaces to be migrated to shared capacity}
}
\value{
A `data.frame` object.
}
\description{
Unassigns the provided workspaces from capacity.Note: The user must have administrator rights (such as Office 365 Global Administrator or Power BI Service Administrator) to call this API. Required scope: Tenant.Read.All or Tenant.ReadWrite.AllTo set the permissions scope, see Register an app.
}
