% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{clone_report_in_group}
\alias{clone_report_in_group}
\title{Clones the specified report from the specified workspace.If after cloning the report and its dataset reside in two different upgraded workspaces or "My Workspace", a shared dataset will be created in the report's workspace. Reports with live connection will lose the live connection when cloning, and will have a direct binding to the target dataset.Required scope: Content.Create Required permissions: Report - Write permissions. Target dataset (if provided) - Build permissions}
\usage{
clone_report_in_group(token, groupId, reportId)
}
\arguments{
\item{groupId}{The workspace id}

\item{reportId}{The report id}
}
\value{
A `data.frame` object.
}
\description{
Clones the specified report from the specified workspace.If after cloning the report and its dataset reside in two different upgraded workspaces or "My Workspace", a shared dataset will be created in the report's workspace. Reports with live connection will lose the live connection when cloning, and will have a direct binding to the target dataset.Required scope: Content.Create Required permissions: Report - Write permissions. Target dataset (if provided) - Build permissions
}
