% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboards.R
\name{clone_tile}
\alias{clone_tile}
\title{Clones the specified tile from "My Workspace".If target report id and target dataset are not specified, the following can occur:}
\usage{
clone_tile(token, dashboardId, tileId)
}
\arguments{
\item{dashboardId}{The dashboard id}

\item{tileId}{The tile id}
}
\value{
A `data.frame` object.
}
\description{
Clones the specified tile from "My Workspace".If target report id and target dataset are not specified, the following can occur:
}
