% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboards.R
\name{clone_tile_in_group}
\alias{clone_tile_in_group}
\title{Clones the specified tile from the specified workspace.If target report id and target dataset are missing, the following can occur:}
\usage{
clone_tile_in_group(token, dashboardId, groupId, tileId)
}
\arguments{
\item{dashboardId}{The dashboard id}

\item{groupId}{The workspace id}

\item{tileId}{The tile id}
}
\value{
A `data.frame` object.
}
\description{
Clones the specified tile from the specified workspace.If target report id and target dataset are missing, the following can occur:
}
