% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateways.R
\name{create_datasource}
\alias{create_datasource}
\title{Creates a new datasource on the specified gateway.Required scope: Dataset.ReadWrite.All}
\usage{
create_datasource(token, gatewayId)
}
\arguments{
\item{gatewayId}{The gateway id}
}
\value{
A `data.frame` object.
}
\description{
Creates a new datasource on the specified gateway.Required scope: Dataset.ReadWrite.All
}
