% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{create_group}
\alias{create_group}
\title{Creates new workspace.Required scope: Workspace.ReadWrite.AllTo set the permissions scope, see Register an app.}
\usage{
create_group(token, workspaceV2)
}
\arguments{
\item{workspaceV2}{Preview feature: Create a workspace V2. The only supported value is true.}
}
\value{
A `data.frame` object.
}
\description{
Creates new workspace.Required scope: Workspace.ReadWrite.AllTo set the permissions scope, see Register an app.
}
