% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{create_temporary_upload_location}
\alias{create_temporary_upload_location}
\title{Creates a temporary blob storage to be used to import large .pbix files larger than 1 GB and up to 10 GB.
To import large .pbix files, create a temporary upload location and upload the .pbix file using the shared access signature (SAS) url from the response, and then call Post Import and specify 'fileUrl' to be the SAS url in the Request BodyNote: Import large .pbix files is only available for workspaces on premium capacity and for .pbix files that are between 1 GB and 10 GB. Required scope: Dataset.ReadWrite.All}
\usage{
create_temporary_upload_location(token)
}
\arguments{
\item{OK}{}
}
\value{
A `data.frame` object.
}
\description{
Creates a temporary blob storage to be used to import large .pbix files larger than 1 GB and up to 10 GB.
To import large .pbix files, create a temporary upload location and upload the .pbix file using the shared access signature (SAS) url from the response, and then call Post Import and specify 'fileUrl' to be the SAS url in the Request BodyNote: Import large .pbix files is only available for workspaces on premium capacity and for .pbix files that are between 1 GB and 10 GB. Required scope: Dataset.ReadWrite.All
}
