% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed_token.R
\name{dashboards_generatetokeningroup}
\alias{dashboards_generatetokeningroup}
\title{Generates an embed token to view the specified dashboard from the specified workspace.This API is relevant only to 'App owns data' embed scenario.Required scope: (all of the below)}
\usage{
dashboards_generatetokeningroup(token, dashboardId, groupId)
}
\arguments{
\item{dashboardId}{The dashboard id}

\item{groupId}{The workspace id}
}
\value{
A `data.frame` object.
}
\description{
Generates an embed token to view the specified dashboard from the specified workspace.This API is relevant only to 'App owns data' embed scenario.Required scope: (all of the below)
}
