% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{dashboards_getdashboardsasadmin}
\alias{dashboards_getdashboardsasadmin}
\title{Returns a list of dashboards for the organization.Note: The user must have administrator rights (such as Office 365 Global Administrator or Power BI Service Administrator) to call this API. This API allows 200 requests per hour at maximum. Required scope: Tenant.Read.All or Tenant.ReadWrite.AllTo set the permissions scope, see Register an app.}
\usage{
dashboards_getdashboardsasadmin(token, expand, filter, skip, top)
}
\arguments{
\item{expand}{Expands related entities inline, receives a comma-separated list of data types. Supported: tiles}

\item{filter}{Filters the results, based on a boolean condition}

\item{skip}{Skips the first n results}

\item{top}{Returns only the first n results}
}
\value{
A `data.frame` object.
}
\description{
Returns a list of dashboards for the organization.Note: The user must have administrator rights (such as Office 365 Global Administrator or Power BI Service Administrator) to call this API. This API allows 200 requests per hour at maximum. Required scope: Tenant.Read.All or Tenant.ReadWrite.AllTo set the permissions scope, see Register an app.
}
