% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{dataflows_getdataflowdatasourcesasadmin}
\alias{dataflows_getdataflowdatasourcesasadmin}
\title{Returns a list of datasources for the specified dataflow.Note: The user must have administrator rights (such as Office 365 Global Administrator or Power BI Service Administrator) to call this API. Required scope: Tenant.Read.AllApplication only and delegated permissions are supported.To set the permissions scope, see Register an app.}
\usage{
dataflows_getdataflowdatasourcesasadmin(token, dataflowId)
}
\arguments{
\item{dataflowId}{The dataflow id}
}
\value{
A `data.frame` object.
}
\description{
Returns a list of datasources for the specified dataflow.Note: The user must have administrator rights (such as Office 365 Global Administrator or Power BI Service Administrator) to call this API. Required scope: Tenant.Read.AllApplication only and delegated permissions are supported.To set the permissions scope, see Register an app.
}
