% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{dataflows_getdataflowsasadmin}
\alias{dataflows_getdataflowsasadmin}
\title{Returns a list of dataflows for the organization.Note: The user must have administrator rights (such as Office 365 Global Administrator or Power BI Service Administrator) to call this API. Required scope: Tenant.Read.All or Tenant.ReadWrite.AllTo set the permissions scope, see Register an app.}
\usage{
dataflows_getdataflowsasadmin(
  token,
  filter = NULL,
  skip = NULL,
  top = NULL,
  output = "value"
)
}
\arguments{
\item{groupId}{OK}
}
\value{
A `data.frame` object.
}
\description{
Returns a list of dataflows for the organization.Note: The user must have administrator rights (such as Office 365 Global Administrator or Power BI Service Administrator) to call this API. Required scope: Tenant.Read.All or Tenant.ReadWrite.AllTo set the permissions scope, see Register an app.
}
