% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_datasets.R
\name{datasets_deleterows}
\alias{datasets_deleterows}
\title{Deletes all rows from the specified table within the specified dataset from "My Workspace".Note: this API supports only Push datasets. Required scope: Dataset.ReadWrite.All}
\usage{
datasets_deleterows(token, datasetId, tableName)
}
\arguments{
\item{datasetId}{The dataset id}

\item{tableName}{The table name}
}
\value{
A `data.frame` object.
}
\description{
Deletes all rows from the specified table within the specified dataset from "My Workspace".Note: this API supports only Push datasets. Required scope: Dataset.ReadWrite.All
}
