% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed_token.R
\name{datasets_generatetokeningroup}
\alias{datasets_generatetokeningroup}
\title{Generates an embed token to Embed Q&A based on the specified dataset from the specified workspace.This API is relevant only to 'App owns data' embed scenario.Required scope: Dataset.ReadWrite.All or Dataset.Read.AllWhen using service principal for authentication, refer to Service Principal with Power BI document along with considerations and limitations section.}
\usage{
datasets_generatetokeningroup(token, datasetId, groupId)
}
\arguments{
\item{datasetId}{The dataset id}

\item{groupId}{The workspace id}
}
\value{
A `data.frame` object.
}
\description{
Generates an embed token to Embed Q&A based on the specified dataset from the specified workspace.This API is relevant only to 'App owns data' embed scenario.Required scope: Dataset.ReadWrite.All or Dataset.Read.AllWhen using service principal for authentication, refer to Service Principal with Power BI document along with considerations and limitations section.
}
