% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{datasets_getdatasetsasadmin}
\alias{datasets_getdatasetsasadmin}
\title{Returns a list of datasets for the organization.Note: The user must have administrator rights (such as Office 365 Global Administrator or Power BI Service Administrator) to call this API. This API processes each request for 2 seconds, in the mean time other requests will be queued. Required scope: Tenant.Read.All or Tenant.ReadWrite.AllTo set the permissions scope, see Register an app.}
\usage{
datasets_getdatasetsasadmin(token, filter, skip, top)
}
\arguments{
\item{filter}{Filters the results, based on a boolean condition}

\item{skip}{Skips the first n results}

\item{top}{Returns only the first n results}
}
\value{
A `data.frame` object.
}
\description{
Returns a list of datasets for the organization.Note: The user must have administrator rights (such as Office 365 Global Administrator or Power BI Service Administrator) to call this API. This API processes each request for 2 seconds, in the mean time other requests will be queued. Required scope: Tenant.Read.All or Tenant.ReadWrite.AllTo set the permissions scope, see Register an app.
}
