% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{datasets_getdatasettodataflowslinksingroupasadmin}
\alias{datasets_getdatasettodataflowslinksingroupasadmin}
\title{Returns a list of upstream dataflows for datasets from the specified workspace.Note: The user must have administrator rights (such as Office 365 Global Administrator or Power BI Service Administrator) to call this API. Required scope: Tenant.Read.All or Tenant.ReadWrite.AllTo set the permissions scope, see Register an app.}
\usage{
datasets_getdatasettodataflowslinksingroupasadmin(token, groupId)
}
\arguments{
\item{groupId}{The workspace id}
}
\value{
A `data.frame` object.
}
\description{
Returns a list of upstream dataflows for datasets from the specified workspace.Note: The user must have administrator rights (such as Office 365 Global Administrator or Power BI Service Administrator) to call this API. Required scope: Tenant.Read.All or Tenant.ReadWrite.AllTo set the permissions scope, see Register an app.
}
