% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_datasets.R
\name{datasets_postdataset}
\alias{datasets_postdataset}
\title{Creates a new dataset on "My Workspace".Note: This API supports only Push datasets.Required scope: Dataset.ReadWrite.All}
\usage{
datasets_postdataset(
  token,
  defaultRetentionPolicy = NULL,
  datasources,
  defaultMode,
  name,
  relationships,
  tables,
  output = "value"
)
}
\arguments{
\item{groupId}{OK}
}
\value{
A `data.frame` object.
}
\description{
Creates a new dataset on "My Workspace".Note: This API supports only Push datasets.Required scope: Dataset.ReadWrite.All
}
