% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflows.R
\name{delete_dataflow}
\alias{delete_dataflow}
\title{Deletes a dataflow from the CDS for Analytics storage, including its definition file and actual model.Required scope: Dataflow.ReadWrite.All}
\usage{
delete_dataflow(token, dataflowId, groupId)
}
\arguments{
\item{dataflowId}{The dataflow id}

\item{groupId}{The workspace id}
}
\value{
A `data.frame` object.
}
\description{
Deletes a dataflow from the CDS for Analytics storage, including its definition file and actual model.Required scope: Dataflow.ReadWrite.All
}
