% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{delete_dataset}
\alias{delete_dataset}
\title{Deletes the specified dataset from "My Workspace".Required scope: Dataset.ReadWrite.All}
\usage{
delete_dataset(token, datasetId)
}
\arguments{
\item{datasetId}{The dataset id}
}
\value{
A `data.frame` object.
}
\description{
Deletes the specified dataset from "My Workspace".Required scope: Dataset.ReadWrite.All
}
