% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateways.R
\name{delete_datasource}
\alias{delete_datasource}
\title{Deletes the specified datasource from the specified gateway.Required scope: Dataset.ReadWrite.All}
\usage{
delete_datasource(token, datasourceId, gatewayId, output = "value")
}
\arguments{
\item{groupId}{OK}
}
\value{
A `data.frame` object.
}
\description{
Deletes the specified datasource from the specified gateway.Required scope: Dataset.ReadWrite.All
}
