% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateways.R
\name{delete_datasource_user}
\alias{delete_datasource_user}
\title{Removes the specified user from the specified datasource.Required scope: Dataset.ReadWrite.All}
\usage{
delete_datasource_user(token, datasourceId, emailAdress, gatewayId)
}
\arguments{
\item{datasourceId}{The datasource id}

\item{emailAdress}{The user's email address or the service principal object id}

\item{gatewayId}{The gateway id}
}
\value{
A `data.frame` object.
}
\description{
Removes the specified user from the specified datasource.Required scope: Dataset.ReadWrite.All
}
