% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{delete_group}
\alias{delete_group}
\title{Deletes the specified workspace.Required scope: Workspace.ReadWrite.AllTo set the permissions scope, see Register an app.}
\usage{
delete_group(token, groupId, output = "value")
}
\arguments{
\item{groupId}{OK}
}
\value{
A `data.frame` object.
}
\description{
Deletes the specified workspace.Required scope: Workspace.ReadWrite.AllTo set the permissions scope, see Register an app.
}
