% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{delete_report}
\alias{delete_report}
\title{Deletes the specified report from "My Workspace".Required scope: Report.ReadWrite.All}
\usage{
delete_report(token, reportId)
}
\arguments{
\item{reportId}{The report id}
}
\value{
A `data.frame` object.
}
\description{
Deletes the specified report from "My Workspace".Required scope: Report.ReadWrite.All
}
