% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{delete_report_in_group}
\alias{delete_report_in_group}
\title{Deletes the specified report from the specified workspace.Required scope: Report.ReadWrite.All}
\usage{
delete_report_in_group(token, groupId, reportId)
}
\arguments{
\item{groupId}{The workspace id}

\item{reportId}{The report id}
}
\value{
A `data.frame` object.
}
\description{
Deletes the specified report from the specified workspace.Required scope: Report.ReadWrite.All
}
