% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_env.R
\name{dist_env}
\alias{dist_env}
\title{Environmental distance matrix}
\usage{
dist_env(x)
}
\arguments{
\item{x}{A \code{matrix} of phenotypic values of genotypes in each environment.
The matrix should be of dimensions \code{i} x \code{j} where \code{i} is the
number of genotypes and \code{j} is the number of environments. The matrix
column names should be environment names.}
}
\value{
A \code{dist} matrix of distances between environments. This \code{dist} object
can be used by clustering functions such as \code{\link[stats]{hclust}}.
}
\description{
Calculates the distance between environments based on the observed phenotypes
of genotypes shared between environments.
}
\details{
The distance between environments (\eqn{D_jj}) cannot be less than \code{0}
and cannot exceed \code{4}. A distance close to \code{0} indicates that the relative
performance of genotypes is similar (i.e. the correlation between genotype
performance across environments is \eqn{r_jj = 1}) while a distance
close to \code{2} indicates that the correlation between genotype performance
is close to \eqn{r_jj = 0}. Finally, as \eqn{D_jj} approaches 4, crossover
interactions are abundant and the correlation of genotype performance is close
to \eqn{r_jj = -1}.
}
\examples{
data("barley")

# Convert to useable matrix
barley_mat <- barley \%>\%
  mutate(env = paste(site, year, sep = "_")) \%>\%
  select(variety, env, yield) \%>\%
  spread(env, yield) \%>\%
  select(-1) \%>\%
  as.matrix()

# Calculate distance
barley_dist <- dist_env(barley_mat)

\dontrun{
# Environment dendrogram
barley_clust <- hclust(barley_dist, method = "average")
plot(barley_clust)
}

}

