% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidenceToSAINT.R
\name{evidenceToSAINT}
\alias{evidenceToSAINT}
\title{Create files for SAINT scoring from an evidence file}
\usage{
evidenceToSAINT(evidence, keys, db, prefix = NULL)
}
\arguments{
\item{evidence}{MaxQuant evidence data frame}

\item{keys}{keys data frame where Condition indicates the baits (columns must include Raw.file, Condition, BioReplicate, Experiment, IsotopeLabelType, TestControl, and an optional SAINTGroups column)}

\item{db}{db object created by loadUniProtFASTA}

\item{prefix}{optional prefix for filenames created by this function (default = NULL)}
}
\value{
returns 1 if all files were written successfully
}
\description{
Create files for SAINT scoring from an evidence file
}
\examples{
db <- loadUniProtFASTA("SwissProt.Hsapiens.fasta")
evidence <- read.delim("evidence.txt")
keys <- read.delim("keys.txt")
evidenceToSAINT(evidence = evidence, keys = keys, db = db, prefix = "010121-JJ")
}
