% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{export_report}
\alias{export_report}
\title{Exports the specified report from "My Workspace" to a .pbix file.Required scope: Report.ReadWrite.All or Report.Read.All}
\usage{
export_report(token, reportId)
}
\arguments{
\item{reportId}{The report id}
}
\value{
A `data.frame` object.
}
\description{
Exports the specified report from "My Workspace" to a .pbix file.Required scope: Report.ReadWrite.All or Report.Read.All
}
