% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{export_report_in_group}
\alias{export_report_in_group}
\title{Exports the specified report from the specified workspace to a .pbix file.Required scope: Report.ReadWrite.All or Report.Read.All}
\usage{
export_report_in_group(token, groupId, reportId, output = "value")
}
\arguments{
\item{groupId}{OK}
}
\value{
A `data.frame` object.
}
\description{
Exports the specified report from the specified workspace to a .pbix file.Required scope: Report.ReadWrite.All or Report.Read.All
}
