% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{export_to_file_in_group}
\alias{export_to_file_in_group}
\title{Exports the specified report from the specified workspace to the requested format.Required scope: (all of the below)}
\usage{
export_to_file_in_group(token, groupId, reportId)
}
\arguments{
\item{groupId}{The workspace id}

\item{reportId}{The report id}
}
\value{
A `data.frame` object.
}
\description{
Exports the specified report from the specified workspace to the requested format.Required scope: (all of the below)
}
