% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_leaflet_labels}
\alias{format_leaflet_labels}
\title{Format outlet name and address for use as a leaflet label}
\usage{
format_leaflet_labels(name, address, city, state, zip)
}
\arguments{
\item{name}{Character. Name of the business}

\item{address}{Character. Address of the business}

\item{city}{Character. City name}

\item{state}{Character. State name}

\item{zip}{Character or numeric. Zip code}
}
\value{
An HTML label
}
\description{
Takes name and address and formats it into an HTML label. This is a shortcut
helper function that is used in \code{\link{pbr_me}}
}
