% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grappos_requests.R
\name{format_pbr_url}
\alias{format_pbr_url}
\title{Formats url based on the location provided}
\usage{
format_pbr_url(location, dist = 100, lim = 50, brand_id = 333)
}
\arguments{
\item{location}{Zip code or (portion of) city name}

\item{dist}{Numeric. The distance in miles to search from the location}

\item{lim}{Numeric. The number of results to be returned}

\item{brand_id}{Numeric. The brand ID to be returned}
}
\value{
A url to be passed to \code{\link{pbr_query}}
}
\description{
This function simply readies the url for use in a GET request
}
