% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_mixed.R
\name{fw_mixed}
\alias{fw_mixed}
\title{Mixed Model Finlay-Wilkinson Regression}
\usage{
fw_mixed(y, GEN, ENV)
}
\arguments{
\item{y}{The vector of the response variables}

\item{GEN}{Vector of genotype names}

\item{ENV}{Vector of environmental names}
}
\description{
Fits Finlay-Wilkinson using mixed models
}
\examples{
# Load data
data("gauch.soy")

fw_out <- fw_mixed(y = gauch.soy$yield, GEN = gauch.soy$gen,
                   ENV = gauch.soy$env)

}
