% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed_token.R
\name{generate_token}
\alias{generate_token}
\title{Generates an embed token for multiple reports, datasets and target workspaces. Reports and datasets do not have to be related. The binding of a report to a dataset can be done during embedding. Target workspaces are workspaces where creation of reports is allowed.This API is relevant only to 'App owns data' embed scenario.Required scope:}
\usage{
generate_token(token, datasets, identities, reports, targetWorkspaces)
}
\arguments{
\item{datasets}{List of datasets}

\item{identities}{List of identities to use for RLS rules.}

\item{reports}{List of reports}

\item{targetWorkspaces}{List of target workspaces}
}
\value{
A `data.frame` object.
}
\description{
Generates an embed token for multiple reports, datasets and target workspaces. Reports and datasets do not have to be related. The binding of a report to a dataset can be done during embedding. Target workspaces are workspaces where creation of reports is allowed.This API is relevant only to 'App owns data' embed scenario.Required scope:
}
