% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{geno_mat_ex}
\alias{geno_mat_ex}
\title{Example Genotype Matrix}
\format{An object of class \code{matrix} with 245 rows and 742 columns.}
\usage{
geno_mat_ex
}
\description{
Example Genotype Matrix
}
\keyword{datasets}
