% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{get_activity_events}
\alias{get_activity_events}
\title{Returns a list of audit activity events for a tenant.Note: The user must have administrator rights (such as Office 365 Global Administrator or Power BI Service Administrator) to call this API. This API allows 200 requests per hour at maximum. Required scope: Tenant.Read.All or Tenant.ReadWrite.All. To call this API, provide either a continuation token or both a start and end date time. StartDateTime and EndDateTime must be in the same UTC day.}
\usage{
get_activity_events(
  token,
  filter,
  continuationToken,
  endDateTime,
  startDateTime
)
}
\arguments{
\item{filter}{Filters the results based on a boolean condition, using 'Activity', 'UserId', or both properties. Supports only 'eq' and 'and' operators.}

\item{continuationToken}{Token required to get the next chunk of the result set}

\item{endDateTime}{End date and time of the window for audit event results. Must be in ISO 8601 compliant UTC format.}

\item{startDateTime}{Start date and time of the window for audit event results. Must be in ISO 8601 compliant UTC format.}
}
\value{
A `data.frame` object.
}
\description{
Returns a list of audit activity events for a tenant.Note: The user must have administrator rights (such as Office 365 Global Administrator or Power BI Service Administrator) to call this API. This API allows 200 requests per hour at maximum. Required scope: Tenant.Read.All or Tenant.ReadWrite.All. To call this API, provide either a continuation token or both a start and end date time. StartDateTime and EndDateTime must be in the same UTC day.
}
