% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{get_app}
\alias{get_app}
\title{Returns the specified installed app.Required scope: App.Read.AllTo set the permissions scope, see Register an app.}
\usage{
get_app(token, appId)
}
\arguments{
\item{appId}{The app id}
}
\value{
A `data.frame` object.
}
\description{
Returns the specified installed app.Required scope: App.Read.AllTo set the permissions scope, see Register an app.
}
