% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{get_apps}
\alias{get_apps}
\title{Returns a list of installed apps.Required scope: App.Read.AllTo set the permissions scope, see Register an app.}
\usage{
get_apps(token)
}
\arguments{
\item{OK}{}
}
\value{
A `data.frame` object.
}
\description{
Returns a list of installed apps.Required scope: App.Read.AllTo set the permissions scope, see Register an app.
}
