% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_features.R
\name{get_available_feature_by_name}
\alias{get_available_feature_by_name}
\title{Returns the specified available feature for user by name.Required scope: None}
\usage{
get_available_feature_by_name(token, featureName, output = "value")
}
\arguments{
\item{groupId}{OK}
}
\value{
A `data.frame` object.
}
\description{
Returns the specified available feature for user by name.Required scope: None
}
