% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_features.R
\name{get_available_features}
\alias{get_available_features}
\title{Returns a list of available features for the userRequired scope: None}
\usage{
get_available_features(token)
}
\arguments{
\item{OK}{}
}
\value{
A `data.frame` object.
}
\description{
Returns a list of available features for the userRequired scope: None
}
