% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacities.R
\name{get_capacities}
\alias{get_capacities}
\title{Returns a list of capacities the user has access to.Required scope: Capacity.Read.All or Capacity.ReadWrite.All}
\usage{
get_capacities(token)
}
\arguments{
\item{OK}{}
}
\value{
A `data.frame` object.
}
\description{
Returns a list of capacities the user has access to.Required scope: Capacity.Read.All or Capacity.ReadWrite.All
}
