% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{get_capacities_as_admin}
\alias{get_capacities_as_admin}
\title{Returns a list of capacities for the organization.Note: The user must have administrator rights (such as Office 365 Global Administrator or Power BI Service Administrator) to call this API. Required scope: Tenant.Read.All or Tenant.ReadWrite.AllTo set the permissions scope, see Register an app.}
\usage{
get_capacities_as_admin(token)
}
\arguments{
\item{Expands}{related entities inline}
}
\value{
A `data.frame` object.
}
\description{
Returns a list of capacities for the organization.Note: The user must have administrator rights (such as Office 365 Global Administrator or Power BI Service Administrator) to call this API. Required scope: Tenant.Read.All or Tenant.ReadWrite.AllTo set the permissions scope, see Register an app.
}
