% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboards.R
\name{get_dashboard_in_group}
\alias{get_dashboard_in_group}
\title{Returns the specified dashboard from the specified workspace.Required scope: Dashboard.ReadWrite.All or Dashboard.Read.All}
\usage{
get_dashboard_in_group(token, dashboardId, groupId)
}
\arguments{
\item{dashboardId}{The dashboard id}

\item{groupId}{The workspace id}
}
\value{
A `data.frame` object.
}
\description{
Returns the specified dashboard from the specified workspace.Required scope: Dashboard.ReadWrite.All or Dashboard.Read.All
}
