% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{get_dashboards}
\alias{get_dashboards}
\title{Returns a list of dashboards from the specified app.Required scope: Dashboard.ReadWrite.All or Dashboard.Read.All}
\usage{
get_dashboards(token, appId)
}
\arguments{
\item{appId}{The app id}
}
\value{
A `data.frame` object.
}
\description{
Returns a list of dashboards from the specified app.Required scope: Dashboard.ReadWrite.All or Dashboard.Read.All
}
