% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboards.R
\name{get_dashboards_in_group}
\alias{get_dashboards_in_group}
\title{Returns a list of dashboards from the specified workspace.Required scope: Dashboard.ReadWrite.All or Dashboard.Read.All}
\usage{
get_dashboards_in_group(token, groupId, output = "value")
}
\arguments{
\item{groupId}{OK}
}
\value{
A `data.frame` object.
}
\description{
Returns a list of dashboards from the specified workspace.Required scope: Dashboard.ReadWrite.All or Dashboard.Read.All
}
