% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflows.R
\name{get_dataflow}
\alias{get_dataflow}
\title{Exports the specified dataflow definition to a .json file.Required scope: Dataflow.ReadWrite.All or Dataflow.Read.All}
\usage{
get_dataflow(token, dataflowId, groupId)
}
\arguments{
\item{dataflowId}{The dataflow id}

\item{groupId}{The workspace id}
}
\value{
A `data.frame` object.
}
\description{
Exports the specified dataflow definition to a .json file.Required scope: Dataflow.ReadWrite.All or Dataflow.Read.All
}
