% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflows.R
\name{get_dataflow_data_sources}
\alias{get_dataflow_data_sources}
\title{Returns a list of datasources for the specified dataflow.Required scope: Dataflow.ReadWrite.All or Dataflow.Read.All}
\usage{
get_dataflow_data_sources(token, dataflowId, groupId)
}
\arguments{
\item{dataflowId}{The dataflow id}

\item{groupId}{The workspace id}
}
\value{
A `data.frame` object.
}
\description{
Returns a list of datasources for the specified dataflow.Required scope: Dataflow.ReadWrite.All or Dataflow.Read.All
}
