% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflow_storage_accounts.R
\name{get_dataflow_storage_accounts}
\alias{get_dataflow_storage_accounts}
\title{Returns a list of dataflow storage accounts the user has access to.Required scope: StorageAccount.Read.All or StorageAccount.ReadWrite.All}
\usage{
get_dataflow_storage_accounts(token, output = "value")
}
\arguments{
\item{groupId}{OK}
}
\value{
A `data.frame` object.
}
\description{
Returns a list of dataflow storage accounts the user has access to.Required scope: StorageAccount.Read.All or StorageAccount.ReadWrite.All
}
