% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_dataset}
\alias{get_dataset}
\title{Returns the specified dataset from "My Workspace".Required scope: Dataset.ReadWrite.All or Dataset.Read.All}
\usage{
get_dataset(token, datasetId)
}
\arguments{
\item{datasetId}{The dataset id}
}
\value{
A `data.frame` object.
}
\description{
Returns the specified dataset from "My Workspace".Required scope: Dataset.ReadWrite.All or Dataset.Read.All
}
