% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_dataset_in_group}
\alias{get_dataset_in_group}
\title{Returns the specified dataset from the specified workspace.Required scope: Dataset.ReadWrite.All or Dataset.Read.All}
\usage{
get_dataset_in_group(token, datasetId, groupId)
}
\arguments{
\item{datasetId}{The dataset id}

\item{groupId}{The workspace id}
}
\value{
A `data.frame` object.
}
\description{
Returns the specified dataset from the specified workspace.Required scope: Dataset.ReadWrite.All or Dataset.Read.All
}
