% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_datasets}
\alias{get_datasets}
\title{Returns a list of datasets from "My Workspace".Required scope: Dataset.ReadWrite.All or Dataset.Read.All}
\usage{
get_datasets(token)
}
\arguments{
\item{OK}{}
}
\value{
A `data.frame` object.
}
\description{
Returns a list of datasets from "My Workspace".Required scope: Dataset.ReadWrite.All or Dataset.Read.All
}
