% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateways.R
\name{get_datasource}
\alias{get_datasource}
\title{Returns the specified datasource from the specified gateway.Required scope: Dataset.ReadWrite.All or Dataset.Read.All}
\usage{
get_datasource(token, datasourceId, gatewayId, output = "value")
}
\arguments{
\item{groupId}{OK}
}
\value{
A `data.frame` object.
}
\description{
Returns the specified datasource from the specified gateway.Required scope: Dataset.ReadWrite.All or Dataset.Read.All
}
