% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateways.R
\name{get_datasource_users}
\alias{get_datasource_users}
\title{Returns a list of users who have access to the specified datasource.Required scope: Dataset.ReadWrite.All or Dataset.Read.AllTo set the permissions scope, see Register an app.}
\usage{
get_datasource_users(token, datasourceId, gatewayId)
}
\arguments{
\item{datasourceId}{The datasource id}

\item{gatewayId}{The gateway id}
}
\value{
A `data.frame` object.
}
\description{
Returns a list of users who have access to the specified datasource.Required scope: Dataset.ReadWrite.All or Dataset.Read.AllTo set the permissions scope, see Register an app.
}
