% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R, R/gateways.R, R/reports.R
\name{get_datasources}
\alias{get_datasources}
\title{Returns a list of datasources for the specified dataset from "My Workspace".Required scope: Dataset.ReadWrite.All or Dataset.Read.All}
\usage{
get_datasources(token, reportId)

get_datasources(token, reportId)

get_datasources(token, reportId)
}
\arguments{
\item{reportId}{The report id}

\item{datasetId}{The dataset id}

\item{gatewayId}{The gateway id}
}
\value{
A `data.frame` object.

A `data.frame` object.

A `data.frame` object.
}
\description{
Returns a list of datasources for the specified dataset from "My Workspace".Required scope: Dataset.ReadWrite.All or Dataset.Read.All

Returns a list of datasources from the specified gateway.Required scope: Dataset.ReadWrite.All or Dataset.Read.All

Returns a list of datasources for the specified RDL report from "My Workspace".Required scope: Report.ReadWrite.All or Report.Read.All
}
