% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R, R/reports.R
\name{get_datasources_in_group}
\alias{get_datasources_in_group}
\title{Returns a list of datasources for the specified dataset from the specified workspace.Required scope: Dataset.ReadWrite.All or Dataset.Read.All}
\usage{
get_datasources_in_group(token, groupId, reportId)

get_datasources_in_group(token, groupId, reportId)
}
\arguments{
\item{groupId}{The workspace id}

\item{reportId}{}

\item{datasetId}{}
}
\value{
A `data.frame` object.

A `data.frame` object.
}
\description{
Returns a list of datasources for the specified dataset from the specified workspace.Required scope: Dataset.ReadWrite.All or Dataset.Read.All

Returns a list of datasources for the specified RDL report from the specified workspace.Required scope: Report.ReadWrite.All or Reportt.Read.All
}
