% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_direct_query_refresh_schedule}
\alias{get_direct_query_refresh_schedule}
\title{Returns the refresh schedule of a specified DirectQuery or LiveConnection dataset from "My Workspace".Required scope: Dataset.ReadWrite.All or Dataset.Read.All}
\usage{
get_direct_query_refresh_schedule(token, datasetId)
}
\arguments{
\item{datasetId}{The dataset id}
}
\value{
A `data.frame` object.
}
\description{
Returns the refresh schedule of a specified DirectQuery or LiveConnection dataset from "My Workspace".Required scope: Dataset.ReadWrite.All or Dataset.Read.All
}
