% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_direct_query_refresh_schedule_in_group}
\alias{get_direct_query_refresh_schedule_in_group}
\title{Returns the refresh schedule of a specified DirectQuery or LiveConnection dataset from the specified workspace.Required scope: Dataset.ReadWrite.All or Dataset.Read.All}
\usage{
get_direct_query_refresh_schedule_in_group(
  token,
  datasetId,
  groupId,
  output = "value"
)
}
\arguments{
\item{groupId}{OK}
}
\value{
A `data.frame` object.
}
\description{
Returns the refresh schedule of a specified DirectQuery or LiveConnection dataset from the specified workspace.Required scope: Dataset.ReadWrite.All or Dataset.Read.All
}
