% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{get_export_to_file_status}
\alias{get_export_to_file_status}
\title{Returns the status of the Export to file job from "My Workspace".Required scope: Report.ReadWrite.All or Report.Read.All}
\usage{
get_export_to_file_status(token, exportId, reportId, output = "value")
}
\arguments{
\item{groupId}{OK}
}
\value{
A `data.frame` object.
}
\description{
Returns the status of the Export to file job from "My Workspace".Required scope: Report.ReadWrite.All or Report.Read.All
}
