% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{get_file_of_export_to_file_in_group}
\alias{get_file_of_export_to_file_in_group}
\title{Returns the file of the Export to file job of the specified report from the specified group.Required scope: Report.ReadWrite.All or Report.Read.All}
\usage{
get_file_of_export_to_file_in_group(
  token,
  exportId,
  groupId,
  reportId,
  output = "value"
)
}
\arguments{
\item{groupId}{OK}
}
\value{
A `data.frame` object.
}
\description{
Returns the file of the Export to file job of the specified report from the specified group.Required scope: Report.ReadWrite.All or Report.Read.All
}
