% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_gateway_datasources}
\alias{get_gateway_datasources}
\title{Returns a list of gateway datasources for the specified dataset from "My Workspace".Note: Use the new Dataset - Get Datasources API instead.Required scope: Dataset.ReadWrite.All or Dataset.Read.All}
\usage{
get_gateway_datasources(token, datasetId, output = "value")
}
\arguments{
\item{groupId}{OK}
}
\value{
A `data.frame` object.
}
\description{
Returns a list of gateway datasources for the specified dataset from "My Workspace".Note: Use the new Dataset - Get Datasources API instead.Required scope: Dataset.ReadWrite.All or Dataset.Read.All
}
