% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateways.R
\name{get_gateways}
\alias{get_gateways}
\title{Returns a list of gateways for which the user is an admin.Required scope: Dataset.ReadWrite.All or Dataset.Read.All}
\usage{
get_gateways(token)
}
\arguments{
\item{OK}{}
}
\value{
A `data.frame` object.
}
\description{
Returns a list of gateways for which the user is an admin.Required scope: Dataset.ReadWrite.All or Dataset.Read.All
}
