% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{get_groups}
\alias{get_groups}
\title{Returns a list of workspaces the user has access to.Note: Users that have been recently added to a group may not have their new group immediately available, see Refresh user permissions.Required scope: Workspace.Read.All or Workspace.ReadWrite.AllTo set the permissions scope, see Register an app.}
\usage{
get_groups(token)
}
\arguments{
\item{Filters}{the results, based on a boolean condition}

\item{Skips}{the first n results}

\item{Returns}{only the first n results}
}
\value{
A `data.frame` object.
}
\description{
Returns a list of workspaces the user has access to.Note: Users that have been recently added to a group may not have their new group immediately available, see Refresh user permissions.Required scope: Workspace.Read.All or Workspace.ReadWrite.AllTo set the permissions scope, see Register an app.
}
