% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{get_import}
\alias{get_import}
\title{Returns the specified import from "My Workspace".Required scope: Dataset.ReadWrite.All or Dataset.Read.All}
\usage{
get_import(token, importId, output = "value")
}
\arguments{
\item{groupId}{OK}
}
\value{
A `data.frame` object.
}
\description{
Returns the specified import from "My Workspace".Required scope: Dataset.ReadWrite.All or Dataset.Read.All
}
