% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{get_import_in_group}
\alias{get_import_in_group}
\title{Returns the specified import from the specified workspace.Required scope: Dataset.ReadWrite.All or Dataset.Read.All}
\usage{
get_import_in_group(token, groupId, importId)
}
\arguments{
\item{groupId}{The workspace id}

\item{importId}{The import id}
}
\value{
A `data.frame` object.
}
\description{
Returns the specified import from the specified workspace.Required scope: Dataset.ReadWrite.All or Dataset.Read.All
}
