% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{get_imports}
\alias{get_imports}
\title{Returns a list of imports from "My Workspace".Required scope: Dataset.ReadWrite.All or Dataset.Read.All}
\usage{
get_imports(token, output = "value")
}
\arguments{
\item{groupId}{OK}
}
\value{
A `data.frame` object.
}
\description{
Returns a list of imports from "My Workspace".Required scope: Dataset.ReadWrite.All or Dataset.Read.All
}
