% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{get_imports_in_group}
\alias{get_imports_in_group}
\title{Returns a list of imports from the specified workspace.Required scope: Dataset.ReadWrite.All or Dataset.Read.All}
\usage{
get_imports_in_group(token, groupId)
}
\arguments{
\item{groupId}{The workspace id}
}
\value{
A `data.frame` object.
}
\description{
Returns a list of imports from the specified workspace.Required scope: Dataset.ReadWrite.All or Dataset.Read.All
}
