% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{get_page_in_group}
\alias{get_page_in_group}
\title{Returns the specified page within the specified report from the specified workspace.Required scope: Report.ReadWrite.All or Report.Read.All}
\usage{
get_page_in_group(token, groupId, pageName, reportId, output = "value")
}
\arguments{
\item{groupId}{OK}
}
\value{
A `data.frame` object.
}
\description{
Returns the specified page within the specified report from the specified workspace.Required scope: Report.ReadWrite.All or Report.Read.All
}
