% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_parameters}
\alias{get_parameters}
\title{Returns a list of parameters for the specified dataset from "My Workspace".Required scope: Dataset.ReadWrite.All}
\usage{
get_parameters(token, datasetId)
}
\arguments{
\item{datasetId}{The dataset id}
}
\value{
A `data.frame` object.
}
\description{
Returns a list of parameters for the specified dataset from "My Workspace".Required scope: Dataset.ReadWrite.All
}
