% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_parameters_in_group}
\alias{get_parameters_in_group}
\title{Returns a list of parameters for the specified dataset from the specified workspace.Required scope: Dataset.ReadWrite.All}
\usage{
get_parameters_in_group(token, datasetId, groupId)
}
\arguments{
\item{datasetId}{}

\item{groupId}{The workspace id}
}
\value{
A `data.frame` object.
}
\description{
Returns a list of parameters for the specified dataset from the specified workspace.Required scope: Dataset.ReadWrite.All
}
