% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_refresh_history_in_group}
\alias{get_refresh_history_in_group}
\title{Returns the refresh history of the specified dataset from the specified workspace.Required scope: Dataset.ReadWrite.All or Dataset.Read.All}
\usage{
get_refresh_history_in_group(token, datasetId, groupId, top, output = "value")
}
\arguments{
\item{groupId}{OK}
}
\value{
A `data.frame` object.
}
\description{
Returns the refresh history of the specified dataset from the specified workspace.Required scope: Dataset.ReadWrite.All or Dataset.Read.All
}
