% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_refresh_schedule}
\alias{get_refresh_schedule}
\title{Returns the refresh schedule of the specified dataset from "My Workspace".Required scope: Dataset.ReadWrite.All or Dataset.Read.All}
\usage{
get_refresh_schedule(token, datasetId, output = "value")
}
\arguments{
\item{groupId}{OK}
}
\value{
A `data.frame` object.
}
\description{
Returns the refresh schedule of the specified dataset from "My Workspace".Required scope: Dataset.ReadWrite.All or Dataset.Read.All
}
