% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R, R/reports.R
\name{get_report}
\alias{get_report}
\title{Returns the specified report from the specified app.Required scope: Report.ReadWrite.All or Report.Read.All}
\usage{
get_report(token, reportId)

get_report(token, reportId)
}
\arguments{
\item{reportId}{The report id}

\item{appId}{The app id}
}
\value{
A `data.frame` object.

A `data.frame` object.
}
\description{
Returns the specified report from the specified app.Required scope: Report.ReadWrite.All or Report.Read.All

Returns the specified report from "My Workspace".Required scope: Report.ReadWrite.All or Report.Read.All
}
