% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R, R/dashboards.R
\name{get_tile}
\alias{get_tile}
\title{Returns the specified tile within the specified dashboard from the specified app.Note: All tile types are supported except for "model tiles", which include datasets and live tiles that contain an entire report page. Required scope: Dashboard.ReadWrite.All or Dashboard.Read.All}
\usage{
get_tile(token, dashboardId, tileId)

get_tile(token, dashboardId, tileId)
}
\arguments{
\item{dashboardId}{The dashboard id}

\item{tileId}{The tile id}

\item{appId}{The app id}
}
\value{
A `data.frame` object.

A `data.frame` object.
}
\description{
Returns the specified tile within the specified dashboard from the specified app.Note: All tile types are supported except for "model tiles", which include datasets and live tiles that contain an entire report page. Required scope: Dashboard.ReadWrite.All or Dashboard.Read.All

Returns the specified tile within the specified dashboard from "My Workspace".Note: All tile types are supported except for "model tiles", which include datasets and live tiles that contain an entire report page. Required scope: Dashboard.ReadWrite.All or Dashboard.Read.All
}
