% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboards.R
\name{get_tile_in_group}
\alias{get_tile_in_group}
\title{Returns the specified tile within the specified dashboard from the specified workspace.Note: All tile types are supported except for "model tiles", which include datasets and live tiles that contain an entire report page. Required scope: Dashboard.ReadWrite.All or Dashboard.Read.All}
\usage{
get_tile_in_group(token, dashboardId, groupId, tileId)
}
\arguments{
\item{dashboardId}{The dashboard id}

\item{groupId}{The workspace id}

\item{tileId}{The tile id}
}
\value{
A `data.frame` object.
}
\description{
Returns the specified tile within the specified dashboard from the specified workspace.Note: All tile types are supported except for "model tiles", which include datasets and live tiles that contain an entire report page. Required scope: Dashboard.ReadWrite.All or Dashboard.Read.All
}
