% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R, R/dashboards.R
\name{get_tiles}
\alias{get_tiles}
\title{Returns a list of tiles within the specified dashboard from the specified app.Required scope: Dashboard.ReadWrite.All or Dashboard.Read.All}
\usage{
get_tiles(token, dashboardId, output = "value")

get_tiles(token, dashboardId, output = "value")
}
\arguments{
\item{groupId}{OK}
}
\value{
A `data.frame` object.

A `data.frame` object.
}
\description{
Returns a list of tiles within the specified dashboard from the specified app.Required scope: Dashboard.ReadWrite.All or Dashboard.Read.All

Returns a list of tiles within the specified dashboard from "My Workspace".Note: All tile types are supported except for "model tiles", which include datasets and live tiles that contain an entire report page. Required scope: Dashboard.ReadWrite.All or Dashboard.Read.All
}
