% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflows.R
\name{get_upstream_dataflows_in_group}
\alias{get_upstream_dataflows_in_group}
\title{Returns a list of upstream dataflows for the specified dataflow.Required scope: Dataflow.ReadWrite.All or Dataflow.Read.AllTo set the permissions scope, see Register an app.}
\usage{
get_upstream_dataflows_in_group(token, dataflowId, groupId)
}
\arguments{
\item{dataflowId}{The dataflow id}

\item{groupId}{The workspace id}
}
\value{
A `data.frame` object.
}
\description{
Returns a list of upstream dataflows for the specified dataflow.Required scope: Dataflow.ReadWrite.All or Dataflow.Read.AllTo set the permissions scope, see Register an app.
}
