% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacities.R
\name{groups_capacityassignmentstatus}
\alias{groups_capacityassignmentstatus}
\title{Gets the status of the assignment to capacity operation of the specified workspace.Note: To perform this operation, the user must be admin on the specified workspace. Required scope: Workspace.Read.All and Workspace.ReadWrite.All}
\usage{
groups_capacityassignmentstatus(token, groupId, output = "value")
}
\arguments{
\item{groupId}{OK}
}
\value{
A `data.frame` object.
}
\description{
Gets the status of the assignment to capacity operation of the specified workspace.Note: To perform this operation, the user must be admin on the specified workspace. Required scope: Workspace.Read.All and Workspace.ReadWrite.All
}
