% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacities.R
\name{groups_capacityassignmentstatusmyworkspace}
\alias{groups_capacityassignmentstatusmyworkspace}
\title{Gets the status of "My Workspace" assignment to capacity operation.Note: Required scope: Workspace.Read.All and Workspace.ReadWrite.All}
\usage{
groups_capacityassignmentstatusmyworkspace(token)
}
\arguments{
\item{OK}{}
}
\value{
A `data.frame` object.
}
\description{
Gets the status of "My Workspace" assignment to capacity operation.Note: Required scope: Workspace.Read.All and Workspace.ReadWrite.All
}
