% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas2.R
\name{gwas}
\alias{gwas}
\title{Genomewide Assocation Analysis Through Linear Mixed-Models}
\usage{
gwas(pheno, geno, fixed = NULL, model = c("simple", "K", "Q", "QK", "G",
  "QG"), n.PC = 0, P3D = TRUE, n.core = 1, impute.method = c("mean",
  "EM", "pass"))
}
\arguments{
\item{pheno}{A data.frame of phenotypic data.}

\item{geno}{A data.frame of marker names, positions, and genotypes.}

\item{fixed}{A formula definining the fixed effects in the \code{pheno} input.}

\item{model}{The type of association model to run. Can be one of: \code{"simple"},
\code{"K"}, \code{"Q"}, \code{"QK"}, \code{"G"}, \code{"QG"}, ... . See \emph{Details}
for more information on these models.}

\item{n.PC}{The number of principal components from singular value decomposition
of the \code{K} matrix to use to correct for population structure. If \code{0},
no correction for population structure is performed.}

\item{P3D}{Logical. Population parameters previous determined.}

\item{n.core}{The number of cores to use when calculating marker scores.}

\item{impute.method}{The method by which to impute the marker genotypes.}

\item{test.qxe}{Logical. Should the QxE interaction be tested simultaneously?}
}
\value{
An object of class \code{gwas}, with marker scores for each trait and relevant
summary statistics
}
\description{
Performs a genomewide association analysis for the main effect of QTL. Most
arguments are taken from the \code{\link[GWAS]{rrBLUP}} function.
}
\details{
"P3D" refers to "population parameters previously determined" and was introduced by
Zhang et al. (2010). If P3D = TRUE, the variance components are estimated once,
and the phenotypic variance-covariance matrix is calculated as
\eqn{var(y) = var(u1) + var(u2) + ... + var(e)} or \eqn{var(y) = ZGZ' + WHW' + ... + R},
where G is the variance-covariance matrix of the first random effect, H is the
variance covariance matrix of the second random effect (up to \emph{n} random
effects), and R is the variance-covariance matrix of the


A brief description of model types follows:

\describe{
  \item{simple}{Tests for marker effect without correcting for background polygenic
  effect or for population structure. \cr
  Model: \eqn{y = X\beta + S\alpha + e}}

  \item{K}{Tests for marker effect while correcting for background polygenic
  effect, but not population structure. \cr
  Model: \eqn{y = X\beta + Zu + S\alpha + e}, where \eqn{u ~ N(0, K\sigma^2_G)}}
  \item{Q}{Tests for marker effect while correcting for population structure,
  but not background polygenic effect. \cr
  Model: \eqn{y = X\beta + Qv + S\alpha + e}}

  \item{QK}{Tests for marker effect while correcting for background polygenic
  effect and for population structure. \cr
  Model: \eqn{y = X\beta + Zu + Qv + S\alpha + e}, where \eqn{u ~ N(0, K\sigma^2_G)}}

  \item{G}{Tests for marker effect while correcting for the background polygenic
  effect of chromosomes other than the one on which the marker resides. Does
  not correct for population structure. \cr
  Model: \eqn{y = X\beta + Zu + S\alpha + e}, \cr where \eqn{u ~ N(0, K\sigma^2_G)} and K
  is the relationship matrix obtained using markers on all chromosomes excluding
  the one on which the putative QTL resides.}

  \item{QG}{Tests for marker effect while correcting for the background polygenic
  effect of chromosomes other than the one on which the marker resides. Does
  correct for population structure. \cr
  Model: \eqn{y = X\beta + Zu + Qv + S\alpha + e}, \cr where \eqn{u ~ N(0, K\sigma^2_G)} and K
  is the relationship matrix obtained using markers on all chromosomes excluding
  the one on which the putative QTL resides.}

}
}
\examples{
data("tr_cap_genos_hmp")
data("tr_cap_phenos_met")

# Filter the genotypes
geno <- tr_cap_genos_hmp \%>\%
  select(which(names(.) \%in\% c("marker", "chrom", "pos", unique(tr_cap_phenos_met$line))))

pheno <- tr_cap_phenos_met

\dontrun{}
# Perform association using the simple model versus the Q+K model
gwas_simple_out <- gwas(pheno = pheno, geno = geno, fixed = ~ trial, model = "simple")
gwas_QK_out <- gwas(pheno = pheno, geno = geno, fixed = ~ trial, model = "QK", n.PC = 2)

# Compare with rrBLUP GWAS function for the Q+K model


# Use a subset of the data to fit all models
# Use the first 25 markers of the first two chromosomes
geno <- geno \%>\%
  filter(chrom \%in\% unique(chrom)[1:2]) \%>\%
  group_by(chrom)

# Use phenotypes from the first two trials
pheno <- pheno \%>\%
  filter(trial \%in\% unique(trial)[1:2])

models <- c("simple", "K", "Q", "QK", "G", "QG")

gwas_out <- map(models, ~gwas(pheno = pheno, geno = geno, fixed = ~trial, model = ., n.PC = 2))


}
