% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_support2.R
\name{H_inv}
\alias{H_inv}
\title{Calculate the inverse of the H matrix}
\usage{
H_inv(Vu, Ve, weights, Zlist, Klist)
}
\arguments{
\item{Vu}{Estimated variance of random effects}

\item{Ve}{Estimated variance of residuals}

\item{weights}{Weights for the random effect(s). If the number of random
effects is 1, weight = 1.}

\item{Z}{List of random effects incidence matrices}

\item{K}{List of covariance matrices for random effects}
}
\description{
Calculate the inverse of the H matrix
}
