% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadComparisonMatrix.R
\name{loadComparisonMatrix}
\alias{loadComparisonMatrix}
\title{Title}
\usage{
loadComparisonMatrix(filename)
}
\arguments{
\item{filename}{comparisons file}
}
\value{
a matrix of comparisons compatible with MSstats
}
\description{
Title
}
\examples{
comparisonMatrix <- loadComparisonMatrix("comparisons.txt")
}
