% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFastaDB.R
\name{loadUniProtFASTA}
\alias{loadUniProtFASTA}
\title{Load a fasta-formatted UniProt protein sequence database}
\usage{
loadUniProtFASTA(dbFilenames, getSequences = FALSE)
}
\arguments{
\item{dbFilenames}{list of database file names}

\item{getSequences}{returns a data frame named dbSequences of protein sequences and protein lengths (default = FALSE)}
}
\value{
List of lookup vectors of protein names (getProteinNames), protein descriptions (getProteinDescriptions) and/or data frame of sequences (dbSequences)
}
\description{
Load a fasta-formatted UniProt protein sequence database
}
\examples{
db <- loadUniProtFASTA(dbFilenames = c("SwissProt.Hsapiens.fasta", "Uniprot.Ggallus.fasta"))
getProteinNames <- db[["getProteinNames"]]
getProteinDescriptions <- db[["getProteinDescriptions"]]
proteinName <- getProteinNames[["P11388"]]
proteinDescription <- getProteinDescriptions[["P11388"]]

db <- loadUniProtFASTA("SwissProt.Human.fasta", getSequences = TRUE)
dbSequences <- db[["dbSequences"]]
sequence <- dbSequences[which(dbSequences$Accession == "P11388), "Sequence"]
sequenceLength <- dbSequences[which(dbSequences$Accession == "P11388), "Length"]
}
