% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grappos_requests.R
\name{location_query}
\alias{location_query}
\title{Query information about a location}
\usage{
location_query(location)
}
\arguments{
\item{location}{Zip code or (portion of) city name}
}
\value{
A data.frame with city name, zip code, lat, and long
}
\description{
This function is used to find the information needed for an entire GET
request url based on just the zip code or city name (or regex)
}
