% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr_test.R
\name{lr_test}
\alias{lr_test}
\title{Compare two models with a likelihood ratio test}
\usage{
lr_test(model1, model2)
}
\arguments{
\item{model1}{A fitted model object.}

\item{model2}{See \code{model1}.}
}
\description{
Performs a likelihood ratio test to compare two models
}
