% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd.R
\name{LSD}
\alias{LSD}
\alias{LSD.lmerMod}
\title{Calculate the LSD using a fitted model object}
\usage{
LSD(object, ...)

\method{LSD}{lmerMod}(object, geno.term, env.term = NULL, ge.term = NULL,
  alpha = 0.05)
}
\arguments{
\item{object}{A model object of class \code{lm}, \code{aov}, \code{glm}, or \code{lmer}.}

\item{geno.term}{A character vector giving the model term for genotypes.}

\item{env.term}{A character vector giving the model term for environments. If \code{NULL} (default),
heritability is calculated within one environment. Note that \code{env.term = NULL} will
ignore multiple environments, if present.}

\item{ge.term}{A character vector giving the model term for genotype-by-environment interaction.}

\item{alpha}{The significance level at which to calculate the LSD.}
}
\description{
Calculate the least-significant difference among the means of genotypes from
a fitted model object.
}
\examples{

# Use the gauch.soy dataset
data("gauch.soy")

# Filter
gauch_soy1 <- gauch.soy \%>\%
  group_by(env) \%>\%
  filter(n_distinct(gen, rep) == 28)

# Fit a linear model using lm
mod <- lm(yield ~ gen + env + gen:env, data = gauch_soy1)
# Calculate heritability
LSD(object = mod, geno.term = "gen", env.term = "env", ge.term = "gen:env", alpha = 0.05)

# Fit a mixed-effects model with lmer
mod1 <- lmer(yield ~ (1|gen) + env + (1|gen:env), data = gauch_soy1)
# Calculate heritability
LSD(object = mod1, geno.term = "gen", env.term = "env", ge.term = "gen:env")

}
