% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd_test.R
\name{LSD_test}
\alias{LSD_test}
\title{Perform an LSD test among means}
\usage{
LSD_test(df, LSD, alpha = 0.05, p.adj = c("none", "holm", "hochberg",
  "hommel", "bonferroni", "BH", "BY", "fdr"))
}
\arguments{
\item{df}{A \code{data.frame} with two columns. The first column includes the names
of elements in the group, and the second column is the mean of that element
in the group.}

\item{LSD}{The estimated least-significant difference (i.e. determined by the
\code{\link{pbr}[LSD]} function.}

\item{alpha}{The significance level for hypothesis testing.}

\item{p.adj}{The p-value adjustment method for multiple comparison testing.}
}
\description{
Performs an least-significant difference test among group means, for example
among genotype means.
}
