% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{map_ex}
\alias{map_ex}
\title{Example genetic map}
\format{An object of class \code{data.frame} with 742 rows and 3 columns.}
\usage{
map_ex
}
\description{
Example genetic map
}
\keyword{datasets}
