% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianNormalizeQuant.R
\name{medianNormalizeQuant}
\alias{medianNormalizeQuant}
\title{Median normalize an MSstats quant data frame}
\usage{
medianNormalizeQuant(quant)
}
\arguments{
\item{quant}{MSstats-generated quant data frame}
}
\value{
a quant dataframe that is median normalized
}
\description{
Median normalize an MSstats quant data frame
}
\examples{
evidence <- read.delim(file = "evidence.txt")
keys <- read.delim(file = "keys.txt")
proteinGroups <- read.delim(file = "proteinGroups.txt")
quant <- MaxQtoMSstatsFormat(evidence = evidence, annotation = keys, proteinGroups = proteinGroups)
quantNorm <- medianNormalizeQuant(quant)
}
