% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{milwaukee}
\alias{milwaukee}
\title{Retailers in and around Milwaukee, WI}
\format{
A data.frame with 50 rows and 12 columns
}
\source{
milwaukee <- pbr_me(53210, map = FALSE)
}
\usage{
milwaukee
}
\description{
A dataset containing the names and locations for the 50 closest
retailers closest to ZIP code 53210.
}
\keyword{datasets}
