% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grappos_requests.R
\name{pbr_query}
\alias{pbr_query}
\title{Query for locations that sell cold ones}
\usage{
pbr_query(location, ...)
}
\arguments{
\item{location}{Zip code or (portion of) city name}

\item{...}{Additional arguments to be passed to \code{\link{format_pbr_url}}}
}
\value{
A data.frame of retailers
}
\description{
Query for locations that sell cold ones
}
