% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{phenos_ex}
\alias{phenos_ex}
\title{Example phenotypic data}
\format{An object of class \code{data.frame} with 245 rows and 5 columns.}
\usage{
phenos_ex
}
\description{
Example phenotypic data
}
\keyword{datasets}
