% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_e.R
\name{plot.gwas}
\alias{plot.gwas}
\title{Plot the output from GWAS}
\usage{
\method{plot}{gwas}(x, fdr.level = 0.05)
}
\arguments{
\item{x}{An object of class \code{gwas}.}

\item{fdr.level}{The false discovery rate level.}
}
\description{
Plot the output from GWAS
}
