% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_AMMI.R
\name{plot_AMMI}
\alias{plot_AMMI}
\title{Plot AMMI Output}
\usage{
plot_AMMI(AMMI.out, x = 0, y = 1, show.gen = TRUE, show.env = TRUE)
}
\arguments{
\item{AMMI.out}{The output from the \code{\link[AMMI]{agricolae}} function.}

\item{x}{The variable for the x-axis. Must be an \code{integer} specifying
the interaction principal component to plot. If \code{0}, the phenotypic
mean is plotted.}

\item{y}{The variable for the y-axis. Must be an \code{integer} specifying
the interaction principal component to plot.}

\item{show.gen}{\code{Logical}. Should the genotypes be highlighted with
arrows and text?}

\item{show.env}{\code{Logical}. Should the environments be highlighted with
arrows and text?}
}
\description{
More flexible plotting function for the output of the \code{AMMI} function
in the \code{agricolae} package.
}

