% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_support.R
\name{plot_gwas}
\alias{plot_gwas}
\title{Plot the output of an association analysis}
\usage{
plot_gwas(x, fdr.level = 0.05, type = c("manhattan", "qq"))
}
\arguments{
\item{x}{An object of class \code{gwas} or a list of such objects.}

\item{fdr.level}{The false discovery rate cutoff for declaring adjusted p-values
significant.}

\item{type}{The type of plot to create. Can be \code{"manhattan"} or
\code{"qq"}.}
}
\description{
Plot the output of an association analysis
}
