% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{popvar.barley}
\alias{popvar.barley}
\title{Example data from the \code{PopVar} package.}
\format{An object of class \code{list} with two elements: 1) \code{pheno} - a
\code{data.frame} with 165 rows and 5 columns. The first column contains the
line name, and the remaining four columns contain phenotypic data on four
quantitative traits; 2) \code{geno} - a \code{data.frame} with 742 rows and 168 columns.
The first 3 columns contain information on the name, chromosome, and genetic
map position (respectively) of 742 molecular markers. Remaining columns
contain information on the genotypic states at each of those markers for all
of the barley lines.}
\source{
This data was obtained and lightly edited from the \code{PopVar} package.
}
\usage{
popvar.barley
}
\description{
A \code{list} of phenotypic and molecular marker data on a population of 165
six-row barley lines.
}
\keyword{datasets}
