% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{popvar_data_genos}
\alias{popvar_data_genos}
\title{Example Genotype Matrix}
\format{An object of class \code{matrix} with 245 rows and 742 columns.}
\usage{
popvar_data_genos
}
\description{
Example Genotype Matrix
}
\keyword{datasets}
