% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{popvar_data_map}
\alias{popvar_data_map}
\title{Example genetic map}
\format{An object of class \code{data.frame} with 742 rows and 3 columns.}
\usage{
popvar_data_map
}
\description{
Example genetic map
}
\keyword{datasets}
