% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{popvar_data_phenos}
\alias{popvar_data_phenos}
\title{Example phenotypic data}
\format{An object of class \code{data.frame} with 245 rows and 5 columns.}
\usage{
popvar_data_phenos
}
\description{
Example phenotypic data
}
\keyword{datasets}
