% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{post_import_in_group}
\alias{post_import_in_group}
\title{Creates new content on the specified workspace from .pbix, .json, Excel, Rdl, or file path in OneDrive for Business.Required scope: Dataset.ReadWrite.All}
\usage{
post_import_in_group(
  token,
  groupId,
  nameConflict,
  skipReport,
  datasetDisplayName,
  connectionType,
  filePath,
  fileUrl,
  output = "value"
)
}
\arguments{
\item{groupId}{OK}
}
\value{
A `data.frame` object.
}
\description{
Creates new content on the specified workspace from .pbix, .json, Excel, Rdl, or file path in OneDrive for Business.Required scope: Dataset.ReadWrite.All
}
