% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gei_analysis_functions.R
\name{pull_effects}
\alias{pull_effects}
\alias{pull_effects.lm}
\title{Pull main and interaction effects from a model}
\usage{
pull_effects(object, gen.col = "gen", env.col = "env")

\method{pull_effects}{lm}(object, gen.col = "gen", env.col = "env")
}
\arguments{
\item{object}{A fitted model object.}

\item{gen.col}{The name of the column in \code{data} containing the genotype factor.}

\item{env.col}{The name of the column in \code{data} containing the environment factor.}
}
\description{
Pull main and interaction effects from a model
}
