% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{rebind_report}
\alias{rebind_report}
\title{Rebinds the specified report from "My Workspace" to the requested dataset.If the dataset resides in a different upgraded workspace, a shared dataset will be created in "My Workspace".Reports with live connection will lose the live connection when rebinding, and will have a direct binding to the target dataset.Required scope: Report.ReadWrite.All Required permissions: Report - Write permissions. Target dataset - Build permissions.}
\usage{
rebind_report(token, reportId)
}
\arguments{
\item{reportId}{The report id}
}
\value{
A `data.frame` object.
}
\description{
Rebinds the specified report from "My Workspace" to the requested dataset.If the dataset resides in a different upgraded workspace, a shared dataset will be created in "My Workspace".Reports with live connection will lose the live connection when rebinding, and will have a direct binding to the target dataset.Required scope: Report.ReadWrite.All Required permissions: Report - Write permissions. Target dataset - Build permissions.
}
