% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflows.R
\name{refresh_dataflow}
\alias{refresh_dataflow}
\title{Triggers a refresh for the specified dataflow.Required scope: Dataflow.ReadWrite.All}
\usage{
refresh_dataflow(token, dataflowId, groupId, notifyOption, output = "value")
}
\arguments{
\item{groupId}{OK}
}
\value{
A `data.frame` object.
}
\description{
Triggers a refresh for the specified dataflow.Required scope: Dataflow.ReadWrite.All
}
