% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{refresh_user_permissions}
\alias{refresh_user_permissions}
\title{Refreshes user permissions in Power BI
When a user is granted permissions to a workspace, app, or artifact, it might not be immediately available through API calls. This operation refreshes user permissions and makes sure the user permissions are fully updated.  Required scope: Workspace.Read.All  or Workspace.ReadWrite.AllTo set the permissions scope, see Register an app.}
\usage{
refresh_user_permissions(token, output = "value")
}
\arguments{
\item{groupId}{OK}
}
\value{
A `data.frame` object.
}
\description{
Refreshes user permissions in Power BI
When a user is granted permissions to a workspace, app, or artifact, it might not be immediately available through API calls. This operation refreshes user permissions and makes sure the user permissions are fully updated.  Required scope: Workspace.Read.All  or Workspace.ReadWrite.AllTo set the permissions scope, see Register an app.
}
