% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed_token.R
\name{reports_generatetokenforcreateingroup}
\alias{reports_generatetokenforcreateingroup}
\title{Generates an embed token to allow report creation on the specified workspace based on the specified dataset.This API is relevant only to 'App owns data' embed scenario.Required scope: (all of the below)}
\usage{
reports_generatetokenforcreateingroup(
  token,
  groupId,
  accessLevel,
  allowSaveAs,
  datasetId,
  identities,
  output = "value"
)
}
\arguments{
\item{groupId}{OK}
}
\value{
A `data.frame` object.
}
\description{
Generates an embed token to allow report creation on the specified workspace based on the specified dataset.This API is relevant only to 'App owns data' embed scenario.Required scope: (all of the below)
}
