% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{reports_getreportsingroupasadmin}
\alias{reports_getreportsingroupasadmin}
\title{Returns a list of reports from the specified workspace.Note: The user must have administrator rights (such as Office 365 Global Administrator or Power BI Service Administrator) to call this API. This API allows 200 requests per hour at maximum. Required scope: Tenant.Read.All or Tenant.ReadWrite.AllTo set the permissions scope, see Register an app.}
\usage{
reports_getreportsingroupasadmin(token, groupId)
}
\arguments{
\item{groupId}{Returns only the first n results}
}
\value{
A `data.frame` object.
}
\description{
Returns a list of reports from the specified workspace.Note: The user must have administrator rights (such as Office 365 Global Administrator or Power BI Service Administrator) to call this API. This API allows 200 requests per hour at maximum. Required scope: Tenant.Read.All or Tenant.ReadWrite.AllTo set the permissions scope, see Register an app.
}
