% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_e.R
\name{score_calc}
\alias{score_calc}
\title{Calculate marker scores in association study}
\usage{
score_calc(M_test, P3D, Hinv, y, X, X_fixed, Z, K, qtlxe = TRUE)
}
\arguments{
\item{M_test}{Marker genotype matrix}

\item{P3D}{Logical - are variance components known or should they be estimated per marker?}

\item{Hinv}{The inverse of the H matrix (from the mixed-model output)}

\item{y}{Vector of response values}

\item{X}{Incidence matrix of fixed effects (must be full-rank)}

\item{X_fixed}{Incidence matrix of environment fixed effects (but not including intercept)}

\item{Z}{Incidence matrix of random effects (i.e. genotypes)}

\item{K}{The covariance matrix of random effects}

\item{qtlxe}{Logical - should QTLxE effects be tested?}
}
\description{
Calculate marker scores in association study
}
\details{
Calculate the p-value for each marker in an association study. The main effect
of each marker is tested using an F-test, while QTLxE interaction is tested using
a Wald test.
}
