% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{set_all_dataset_connections_in_group}
\alias{set_all_dataset_connections_in_group}
\title{Note: This API is deprecated and no longer supported.Updates all connections for the specified dataset from the specified workspace.Notes:}
\usage{
set_all_dataset_connections_in_group(token, datasetId, groupId)
}
\arguments{
\item{datasetId}{The dataset id}

\item{groupId}{The workspace id}
}
\value{
A `data.frame` object.
}
\description{
Note: This API is deprecated and no longer supported.Updates all connections for the specified dataset from the specified workspace.Notes:
}
