% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed_token.R
\name{tiles_generatetokeningroup}
\alias{tiles_generatetokeningroup}
\title{Generates an embed token to view the specified tile from the specified workspace.This API is relevant only to 'App owns data' embed scenario.Required scope: (all of the below)}
\usage{
tiles_generatetokeningroup(token, dashboardId, groupId, tileId)
}
\arguments{
\item{dashboardId}{The dashboard id}

\item{groupId}{The workspace id}

\item{tileId}{The tile id}
}
\value{
A `data.frame` object.
}
\description{
Generates an embed token to view the specified tile from the specified workspace.This API is relevant only to 'App owns data' embed scenario.Required scope: (all of the below)
}
