% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{tr_cap_genos}
\alias{tr_cap_genos}
\title{Genotypic data from the two-row barley CAP}
\format{An object of class \code{matrix} with 1617 rows and 2166 columns.}
\usage{
tr_cap_genos
}
\description{
Genotypic data from the two-row barley CAP
}
\keyword{datasets}
