% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{tr_cap_genos_mat}
\alias{tr_cap_genos_mat}
\title{Genotypic data on a two-row barley population}
\format{An object of class \code{matrix} with 1622 rows and 2716 columns. The row names
are the names of barley lines, and the column names are the names of markers.}
\source{
This data was downloaded from the Triticeae Toolbox at https://triticeaetoolbox.org/barley/.
}
\usage{
tr_cap_genos_mat
}
\description{
A \code{matrix} of genomewide marker data. 1622 barley individuals were genotyped
with 2716 DNA markers. The data has been curated and filtered to remove markers
and individuals with a high amount of missing data.
}
\keyword{datasets}
