% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{update_datasources}
\alias{update_datasources}
\title{Updates the datasources of the specified dataset from "My Workspace".Important:}
\usage{
update_datasources(token, datasetId, updateDetails, output = "value")
}
\arguments{
\item{groupId}{OK}
}
\value{
A `data.frame` object.
}
\description{
Updates the datasources of the specified dataset from "My Workspace".Important:
}
