% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{update_group_user}
\alias{update_group_user}
\title{Update the specified user permissions to the specified workspace.Notes:}
\usage{
update_group_user(token, groupId)
}
\arguments{
\item{groupId}{The workspace id}
}
\value{
A `data.frame` object.
}
\description{
Update the specified user permissions to the specified workspace.Notes:
}
