% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{update_parameters_in_group}
\alias{update_parameters_in_group}
\title{Updates the parameters values for the specified dataset from the specified workspace.Important: The dataset must be refreshed for the new parameter values to be applied. Wait 30 minutes for the update parameters operation to complete before refreshing.Required scope: Dataset.ReadWrite.All}
\usage{
update_parameters_in_group(token, datasetId, groupId)
}
\arguments{
\item{datasetId}{}

\item{groupId}{The workspace id}
}
\value{
A `data.frame` object.
}
\description{
Updates the parameters values for the specified dataset from the specified workspace.Important: The dataset must be refreshed for the new parameter values to be applied. Wait 30 minutes for the update parameters operation to complete before refreshing.Required scope: Dataset.ReadWrite.All
}
