% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflows.R, R/datasets.R
\name{update_refresh_schedule}
\alias{update_refresh_schedule}
\title{Creates or updates the specified dataflow refresh schedule configuration.Required scope: Dataflow.ReadWrite.All}
\usage{
update_refresh_schedule(token, datasetId, value, output = "value")

update_refresh_schedule(token, datasetId, value, output = "value")
}
\arguments{
\item{groupId}{OK}
}
\value{
A `data.frame` object.

A `data.frame` object.
}
\description{
Creates or updates the specified dataflow refresh schedule configuration.Required scope: Dataflow.ReadWrite.All

Updates the refresh schedule for the specified dataset from "My Workspace".This operation is only supported for the dataset owner.A request that disables the refresh schedule should contain no other changes.The days array should not be set to empty array.The times may be set to empty array (in which case Power BI will use a default single time per day).The limit on number of time slots per day depends on the type of capacity used (Premium or Shared), see What is Microsoft Power BI Premium.Required scope: Dataset.ReadWrite.All
}
