% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{update_report_content}
\alias{update_report_content}
\title{Updates the specified report from "My Workspace" to have the same content as the report in the request body.Required scope: Report.ReadWrite.All}
\usage{
update_report_content(
  token,
  reportId,
  sourceReport,
  sourceType,
  output = "value"
)
}
\arguments{
\item{groupId}{OK}
}
\value{
A `data.frame` object.
}
\description{
Updates the specified report from "My Workspace" to have the same content as the report in the request body.Required scope: Report.ReadWrite.All
}
