% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{update_report_content_in_group}
\alias{update_report_content_in_group}
\title{Updates the specified report from the specified workspace to have the same content as the specified report in the request body.Required scope: Report.ReadWrite.All}
\usage{
update_report_content_in_group(token, groupId, reportId)
}
\arguments{
\item{groupId}{The workspace id}

\item{reportId}{The report id}
}
\value{
A `data.frame` object.
}
\description{
Updates the specified report from the specified workspace to have the same content as the specified report in the request body.Required scope: Report.ReadWrite.All
}
