% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboards.R
\name{add_dashboard}
\alias{add_dashboard}
\title{Creates a new empty dashboard on "My Workspace".Required scope: Content.Create}
\usage{
add_dashboard(token, name, output = "value")
}
\arguments{
\item{groupId}{OK}
}
\value{
A `data.frame` object.
}
\description{
Creates a new empty dashboard on "My Workspace".Required scope: Content.Create
}
