% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_datasets.R
\name{datasets_postdatasetingroup}
\alias{datasets_postdatasetingroup}
\title{Creates a new dataset in the specified workspace.Note: This API supports only Push datasets.Required scope: Dataset.ReadWrite.All}
\usage{
datasets_postdatasetingroup(
  token,
  groupId,
  defaultRetentionPolicy,
  datasources,
  defaultMode,
  name,
  relationships,
  tables,
  output = "value"
)
}
\arguments{
\item{groupId}{OK}
}
\value{
A `data.frame` object.
}
\description{
Creates a new dataset in the specified workspace.Note: This API supports only Push datasets.Required scope: Dataset.ReadWrite.All
}
