% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_datasets.R
\name{datasets_postrowsingroup}
\alias{datasets_postrowsingroup}
\title{Adds new data rows to the specified table, within the specified dataset, from the specified workspace.Note: This API supports only Push datasets. REST API Limitations: See Power BI REST API limitations. Required scope: Dataset.ReadWrite.All}
\usage{
datasets_postrowsingroup(token, datasetId, groupId, tableName)
}
\arguments{
\item{datasetId}{The dataset id}

\item{groupId}{The workspace id}

\item{tableName}{The table name}
}
\value{
A `data.frame` object.
}
\description{
Adds new data rows to the specified table, within the specified dataset, from the specified workspace.Note: This API supports only Push datasets. REST API Limitations: See Power BI REST API limitations. Required scope: Dataset.ReadWrite.All
}
