% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{delete_dataset_in_group}
\alias{delete_dataset_in_group}
\title{Deletes the specified dataset from the specified workspace.Required scope: Dataset.ReadWrite.All}
\usage{
delete_dataset_in_group(token, datasetId, groupId)
}
\arguments{
\item{datasetId}{The dataset id}

\item{groupId}{The workspace id}
}
\value{
A `data.frame` object.
}
\description{
Deletes the specified dataset from the specified workspace.Required scope: Dataset.ReadWrite.All
}
