% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{delete_user_in_group}
\alias{delete_user_in_group}
\title{Deletes the specified user permissions from the specified workspace.Required scope: Workspace.ReadWrite.AllTo set the permissions scope, see Register an app.}
\usage{
delete_user_in_group(token, groupId, user, output = "value")
}
\arguments{
\item{groupId}{OK}
}
\value{
A `data.frame` object.
}
\description{
Deletes the specified user permissions from the specified workspace.Required scope: Workspace.ReadWrite.AllTo set the permissions scope, see Register an app.
}
