% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_env.R
\name{dist_env}
\alias{dist_env}
\title{Environmental distance matrix}
\usage{
dist_env(x, gen.col = "gen", env.col = "env", pheno.col = "yield")
}
\arguments{
\item{x}{A \code{data.frame} of phenotypic observations on genotypes in environments. The data.frame
should be in tidy format, where every phenotypic observation has a separate row.}

\item{gen.col}{The column name that contains the names of genotypes.}

\item{env.col}{The column name that contains the names of environments.}

\item{pheno.col}{The column name that contains the phenotypic observations.}
}
\value{
A \code{dist} matrix of distances between environments. This \code{dist} object
can be used by clustering functions such as \code{\link[stats]{hclust}}.
}
\description{
Calculates the distance between environments based on the observed phenotypes
of genotypes shared between environments.
}
\details{
The distance between environments (\eqn{D_jj}) cannot be less than \code{0}
and cannot exceed \code{4}. A distance close to \code{0} indicates that the relative
performance of genotypes is similar (i.e. the correlation between genotype
performance across environments is \eqn{r_jj = 1}) while a distance
close to \code{2} indicates that the correlation between genotype performance
is close to \eqn{r_jj = 0}. Finally, as \eqn{D_jj} approaches 4, crossover
interactions are abundant and the correlation of genotype performance is close
to \eqn{r_jj = -1}.
}
\examples{

\dontrun{
data("yang.barley")

# Calculate distance
yang.barley.dist <- dist_env(x = yang.barley, env.col = "site")

# Environment dendrogram
yang.barley.clust <- hclust(yang.barley.dist, method = "average")
plot(yang.barley.clust)
}

}

