% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{export_to_file}
\alias{export_to_file}
\title{Exports the specified report from "My Workspace" to the requested format.Required scope: (all of the below)}
\usage{
export_to_file(token, reportId)
}
\arguments{
\item{reportId}{The report id}
}
\value{
A `data.frame` object.
}
\description{
Exports the specified report from "My Workspace" to the requested format.Required scope: (all of the below)
}
