% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R, R/dashboards.R
\name{get_dashboards}
\alias{get_dashboards}
\title{Returns a list of dashboards from the specified app.Required scope: Dashboard.ReadWrite.All or Dashboard.Read.All}
\usage{
get_dashboards(token)

get_dashboards(token)
}
\arguments{
\item{appId}{The app id}

\item{OK}{}
}
\value{
A `data.frame` object.

A `data.frame` object.
}
\description{
Returns a list of dashboards from the specified app.Required scope: Dashboard.ReadWrite.All or Dashboard.Read.All

Returns a list of dashboards from "My Workspace".Required scope: Dashboard.ReadWrite.All or Dashboard.Read.All
}
