% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflows.R
\name{get_dataflows}
\alias{get_dataflows}
\title{Returns a list of all dataflows from the specified workspace.Required scope: Dataflow.ReadWrite.All or Dataflow.Read.All}
\usage{
get_dataflows(token, groupId)
}
\arguments{
\item{groupId}{The workspace id}
}
\value{
A `data.frame` object.
}
\description{
Returns a list of all dataflows from the specified workspace.Required scope: Dataflow.ReadWrite.All or Dataflow.Read.All
}
