% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{get_export_to_file_status_in_group}
\alias{get_export_to_file_status_in_group}
\title{Returns the status of the Export to file job from the specified workspace.Required scope: Report.ReadWrite.All or Report.Read.All}
\usage{
get_export_to_file_status_in_group(token, exportId, groupId, reportId)
}
\arguments{
\item{exportId}{The export id}

\item{groupId}{The workspace id}

\item{reportId}{The report id}
}
\value{
A `data.frame` object.
}
\description{
Returns the status of the Export to file job from the specified workspace.Required scope: Report.ReadWrite.All or Report.Read.All
}
