% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{get_file_of_export_to_file}
\alias{get_file_of_export_to_file}
\title{Returns the file of the Export to file job of the specified report from "My Workspace".Required scope: Report.ReadWrite.All or Report.Read.All}
\usage{
get_file_of_export_to_file(token, exportId, reportId)
}
\arguments{
\item{exportId}{The export id}

\item{reportId}{The report id}
}
\value{
A `data.frame` object.
}
\description{
Returns the file of the Export to file job of the specified report from "My Workspace".Required scope: Report.ReadWrite.All or Report.Read.All
}
