% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateways.R
\name{get_gateway}
\alias{get_gateway}
\title{Returns the specified gateway.Required scope: Dataset.ReadWrite.All or Dataset.Read.All}
\usage{
get_gateway(token, gatewayId)
}
\arguments{
\item{gatewayId}{The gateway id}
}
\value{
A `data.frame` object.
}
\description{
Returns the specified gateway.Required scope: Dataset.ReadWrite.All or Dataset.Read.All
}
