% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{get_pages}
\alias{get_pages}
\title{Returns a list of pages within the specified report from "My Workspace".Required scope: Report.ReadWrite.All or Report.Read.All}
\usage{
get_pages(token, reportId)
}
\arguments{
\item{reportId}{The report id}
}
\value{
A `data.frame` object.
}
\description{
Returns a list of pages within the specified report from "My Workspace".Required scope: Report.ReadWrite.All or Report.Read.All
}
