% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_refresh_history_in_group}
\alias{get_refresh_history_in_group}
\title{Returns the refresh history of the specified dataset from the specified workspace.Required scope: Dataset.ReadWrite.All or Dataset.Read.All}
\usage{
get_refresh_history_in_group(token, datasetId, groupId, top)
}
\arguments{
\item{datasetId}{The dataset id}

\item{groupId}{The workspace id}

\item{top}{The requested number of entries in the refresh history. If not provided, the default is all available entries.}
}
\value{
A `data.frame` object.
}
\description{
Returns the refresh history of the specified dataset from the specified workspace.Required scope: Dataset.ReadWrite.All or Dataset.Read.All
}
