% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{get_reports}
\alias{get_reports}
\title{Returns a list of reports from the specified app.Required scope: Report.ReadWrite.All or Report.Read.All}
\usage{
get_reports(token, appId)
}
\arguments{
\item{appId}{The app id}
}
\value{
A `data.frame` object.
}
\description{
Returns a list of reports from the specified app.Required scope: Report.ReadWrite.All or Report.Read.All
}
