% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_e.R
\name{gwas_e}
\alias{gwas_e}
\title{Genomewide Association Analysis for Multiple Enviroments}
\usage{
gwas_e(pheno, geno, fixed = NULL, K = NULL, n.PC = 0, P3D = TRUE,
  n.core = 1, impute.method = c("mean", "EM", "pass"))
}
\arguments{
\item{pheno}{A data.frame of phenotypic data.}

\item{geno}{A data.frame of marker names, positions, and genotypes.}

\item{fixed}{A character vector of fixed effects.}

\item{K}{The covariance matrix of random effects (i.e. genotypes).}

\item{n.PC}{The number of principal components from singular value decomposition
of the \code{K} matrix to use to correct for population structure.}

\item{P3D}{Logical. Population parameters previous determined.}

\item{n.core}{The number of cores to use when calculating marker scores.}

\item{impute.method}{The method by which to impute the marker genotypes.}
}
\value{
An object of class \code{gwas}, with marker scores for each trait.
}
\description{
Performs a genomewide association analysis for the main effect of markers and
their interaction with the environment. Most arguments are taken from the
\code{\link[GWAS]{rrBLUP}} function.
}
\examples{

data("tr_cap_genos_hmp")
data("tr_cap_phenos_met")

# Filter the genotypes
geno <- subset(tr_cap_genos_hmp,
               select = which(names(tr_cap_genos_hmp) \%in\%
                          c("marker", "chrom", "pos", unique(tr_cap_phenos_met$line))))

pheno <- tr_cap_phenos_met

\dontrun{}
# Perform association
gwas_out <- gwas_e(pheno = pheno, geno = geno, fixed = "trial")

}
