% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_support.R
\name{H_inv}
\alias{H_inv}
\title{Calculate the inverse of the H matrix}
\usage{
H_inv(Vu, Ve, n, Z, K)
}
\arguments{
\item{Vu}{Estimated variance of random effects}

\item{Ve}{Estimated variance of residuals}

\item{n}{Number of observations}

\item{Z}{Random effects incidence matrix}

\item{K}{Covariance matrix for random effects}
}
\description{
Calculate the inverse of the H matrix
}
