% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{harm_mean}
\alias{harm_mean}
\title{Convenience functions}
\usage{
harm_mean(x)
}
\description{
Convenience functions. These are generally not to be called by the user
}
\details{
The formula for the harmonic mean given in the function \code{harm_mean} is:

\eqn{e_h = n / \sum{\frac{1}{e_j}}}
}
