% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herit_boot.R, R/herit_boot_methods.R
\name{herit_boot}
\alias{herit_boot}
\alias{herit_boot.lm}
\alias{herit_boot.lmerMod}
\title{Heritability confidence interval by bootstrap}
\usage{
herit_boot(object, ...)

\method{herit_boot}{lm}(object, geno.term, env.term = NULL, ge.term = NULL,
  boot.reps = 1000, alpha = 0.95)

\method{herit_boot}{lmerMod}(object, geno.term, env.term = NULL,
  ge.term = NULL, boot.reps = 1000, alpha = 0.95)
}
\arguments{
\item{object}{A model object of class \code{lm}, \code{aov}, \code{glm}, or \code{lmer}.}

\item{geno.term}{A character vector giving the model term for genotypes.}

\item{env.term}{A character vector giving the model term for environments. If \code{NULL} (default),
heritability is calculated within one environment. Note that \code{env.term = NULL} will
ignore multiple environments, if present.}

\item{ge.term}{A character vector giving the model term for genotype-by-environment interaction.}

\item{boot.reps}{The number of bootsrap replicates.}

\item{alpha}{The significance level for the confidence interval.}
}
\description{
Heritability confidence interval by bootstrap
}
\section{Methods (by class)}{
\itemize{
\item \code{lm}: 

\item \code{lmerMod}: 
}}

