% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herit_boot.R
\name{herit_boot}
\alias{herit_boot}
\alias{herit_boot.lmerMod}
\title{Heritability confidence interval by bootstrap}
\usage{
herit_boot(object, exp, boot.reps = 1000, ...)

\method{herit_boot}{lmerMod}(object, exp, boot.reps = 1000, alpha = 0.05,
  ...)
}
\arguments{
\item{object}{A model object. See \emph{Details} for accepted classes.}

\item{exp}{A quoted expression used to calculate the heritability. For instance,
\code{"geno / (geno + (Residual / r))"}.}

\item{boot.reps}{The number of bootsrap replicates.}

\item{...}{Other arguments to pass. This is generally a list of other objects
that are in \code{exp}, but may not be found in \code{object}. For instance, if
the argument \code{exp = "geno / (geno + (Residual / r))"} was passed, you would
also pass the argument \code{r = 2}.}

\item{alpha}{The significance level for the confidence interval.}
}
\value{
A \code{data.frame} with the following values:

\describe{
  \item{heritability}{The estimate of the heritability using the original data.}
  \item{se}{The standard error of the estimate, calculated as the standard deviation
  among the bootstrap replicates.}
  \item{bias}{The bias of the original heritability estimate, calculated as the difference
  between the mean of the bootstrapped estimates and the original estimate.}
  \item{ci_lower}{The lower limit of the confidence interval.}
  \item{ci_upper}{The upper limit of the confidence interval.}
}
}
\description{
Heritability confidence interval by bootstrap
}
\details{
This function implements model-based bootstrapping to obtain the standard error
and confidence interval around an estimate of the heritability. The function
uses the original fitted model to simulate new observations using the same
model parameters. The model is then re-fitted using the new observations,
and the heritability is re-calculated. This is repeated \emph{n} times.
}
