% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacities.R
\name{patch_workload}
\alias{patch_workload}
\title{Changes the state of a specific workload to Enabled or Disabled. When enabling a workload the maximum memory percentage that the workload can consume must be set.Required scope: Capacity.ReadWrite.All}
\usage{
patch_workload(
  token,
  capacityId,
  workloadName,
  maxMemoryPercentageSetByUser,
  state,
  output = "value"
)
}
\arguments{
\item{groupId}{OK}
}
\value{
A `data.frame` object.
}
\description{
Changes the state of a specific workload to Enabled or Disabled. When enabling a workload the maximum memory percentage that the workload can consume must be set.Required scope: Capacity.ReadWrite.All
}
