% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar.R
\name{pbr}
\alias{pbr}
\title{Progress bar}
\usage{
pbr(iterations, width = NULL, min_width = 10, max_width = 100L)
}
\arguments{
\item{iterations}{Number of total expected iterations}

\item{width}{Desired width of the progress bar (including text)}

\item{min_width}{Desired minimum width}

\item{max_width}{Desired max width}
}
\value{
A progress bar list with tick and done functions
}
\description{
Create a progress bar object
}
