% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbr.R
\name{pbr}
\alias{pbr}
\title{Progress bar}
\usage{
pbr(iterations = 10, width = NULL, min_width = 10, max_width = 1000L)
}
\arguments{
\item{iterations}{Number of total expected iterations}

\item{width}{Desired width of the progress bar (including text)}

\item{min_width}{Desired minimum width}

\item{max_width}{Desired max width}
}
\value{
A progress bar list with tick and done functions
}
\description{
Create a progress bar object
}
\examples{
## define test function with a for loop
pb_test <- function(n = 50, secs = 3) {
  pb <- pbr(n)
  on.exit(pb$done(), add = TRUE)
  for (i in seq_len(n)) {
    Sys.sleep(secs / n)
    pb$tick()
  }
}
## execute test function
pb_test()
}
