% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{post_import}
\alias{post_import}
\title{Creates new content on "My Workspace" from .pbix, Excel, Rdl or file path in OneDrive for Business.Required scope: Dataset.ReadWrite.All}
\usage{
post_import(
  token,
  nameConflict,
  skipReport,
  datasetDisplayName,
  connectionType,
  filePath,
  fileUrl,
  output = "value"
)
}
\arguments{
\item{groupId}{OK}
}
\value{
A `data.frame` object.
}
\description{
Creates new content on "My Workspace" from .pbix, Excel, Rdl or file path in OneDrive for Business.Required scope: Dataset.ReadWrite.All
}
