% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{post_import_in_group}
\alias{post_import_in_group}
\title{Creates new content on the specified workspace from .pbix, .json, Excel, Rdl, or file path in OneDrive for Business.Required scope: Dataset.ReadWrite.All}
\usage{
post_import_in_group(
  token,
  groupId,
  nameConflict,
  skipReport,
  datasetDisplayName
)
}
\arguments{
\item{groupId}{The workspace id}

\item{nameConflict}{Determines what to do if a dataset with the same name already exists. Default value is 'Ignore'.Only Abort and Overwrite are supported with Rdl files.Only Abort and GenerateUniqueName are supported with dataflow model.json files.}

\item{skipReport}{Determines whether to skip report import, if specified value must be 'true'. Only supported for PBIX files.}

\item{datasetDisplayName}{The display name of the dataset should include file extension. Not supported when importing from OneDrive for Business. For importing or creating dataflows, this parameter should be hardcoded to model.json.}
}
\value{
A `data.frame` object.
}
\description{
Creates new content on the specified workspace from .pbix, .json, Excel, Rdl, or file path in OneDrive for Business.Required scope: Dataset.ReadWrite.All
}
