% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed_token.R
\name{reports_generatetokeningroup}
\alias{reports_generatetokeningroup}
\title{Generates an embed token to view or edit the specified report from the specified workspace.This API is relevant only to 'App owns data' embed scenario.Required scope: (all of the below)}
\usage{
reports_generatetokeningroup(
  token,
  groupId,
  reportId,
  accessLevel,
  allowSaveAs,
  datasetId,
  identities,
  output = "value"
)
}
\arguments{
\item{groupId}{OK}
}
\value{
A `data.frame` object.
}
\description{
Generates an embed token to view or edit the specified report from the specified workspace.This API is relevant only to 'App owns data' embed scenario.Required scope: (all of the below)
}
