% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_support.R
\name{score_calc}
\alias{score_calc}
\title{Calculate marker scores in association study}
\usage{
score_calc(M_test, model, snp_info, P3D, Hinv, y, X, Z0, K0, X_fixed, Z_rand)
}
\arguments{
\item{M_test}{Marker genotype matrix}

\item{model}{The model name.}

\item{snp_info}{A data.frame of marker names, chromosomes, and postitions. This
is used for models "G" and "QG".}

\item{P3D}{Logical - are variance components known or should they be estimated per marker?}

\item{Hinv}{List of inverse of the phenotype vcov matrix (from the mixed-model output).
Must be of length 1 or the number of chromosomes.}

\item{y}{Vector of response values}

\item{X}{List of incidence matrices of fixed effects (must be full-rank, i.e. from svd).
Must of length 1 or the number of chromosomes.}

\item{Z0}{Incidence matrix of random main effects (i.e. genotypes)}

\item{K0}{List of covariance matrices of random main effects. Must be of length
1 or the number of chromosomes.}

\item{X_fixed}{Incidence matrix of environment fixed effects (but not including intercept)}

\item{Z_rand}{Incidence matrix of genotype random effects (but not including intercept)}
}
\description{
Calculate marker scores in association study
}
\details{
Calculate the p-value for each marker in an association study. The main effect
of each marker is tested using a Wald test.
}
