% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{set_all_dataset_connections}
\alias{set_all_dataset_connections}
\title{Note: This API is deprecated and no longer supported.Updates all connections for the specified dataset from "My Workspace".Notes:}
\usage{
set_all_dataset_connections(
  token,
  datasetId,
  connectionString,
  output = "value"
)
}
\arguments{
\item{groupId}{OK}
}
\value{
A `data.frame` object.
}
\description{
Note: This API is deprecated and no longer supported.Updates all connections for the specified dataset from "My Workspace".Notes:
}
