% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{take_over_in_group}
\alias{take_over_in_group}
\title{Transfers ownership over the specified dataset to the current authorized user.Required scope: Dataset.ReadWrite.All}
\usage{
take_over_in_group(token, datasetId, groupId)
}
\arguments{
\item{datasetId}{The dataset id}

\item{groupId}{The workspace id}
}
\value{
A `data.frame` object.
}
\description{
Transfers ownership over the specified dataset to the current authorized user.Required scope: Dataset.ReadWrite.All
}
