% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{tr_cap_genos_hmp}
\alias{tr_cap_genos_hmp}
\title{Marker genetic map for the two-row barley CAP}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2187 rows and 1618 columns.}
\usage{
tr_cap_genos_hmp
}
\description{
Marker genetic map for the two-row barley CAP
}
\keyword{datasets}
