% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{tr_cap_genos_mat}
\alias{tr_cap_genos_mat}
\title{Genotypic data from the two-row barley CAP}
\format{An object of class \code{matrix} with 1614 rows and 2187 columns.}
\usage{
tr_cap_genos_mat
}
\description{
Genotypic data from the two-row barley CAP
}
\keyword{datasets}
