% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{tr_cap_phenos}
\alias{tr_cap_phenos}
\title{Phenotypic data for the two-row barley CAP}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1617 rows and 5 columns.}
\usage{
tr_cap_phenos
}
\description{
Phenotypic data for the two-row barley CAP
}
\keyword{datasets}
