% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateways.R
\name{update_datasource}
\alias{update_datasource}
\title{Updates the credentials of the specified datasource from the specified gateway.Required scope: Dataset.ReadWrite.All}
\usage{
update_datasource(token, datasourceId, gatewayId)
}
\arguments{
\item{datasourceId}{The datasource id}

\item{gatewayId}{The gateway id}
}
\value{
A `data.frame` object.
}
\description{
Updates the credentials of the specified datasource from the specified gateway.Required scope: Dataset.ReadWrite.All
}
