% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{update_datasources_in_group}
\alias{update_datasources_in_group}
\title{Updates the datasources of the specified dataset from the specified workspace.Important:}
\usage{
update_datasources_in_group(token, datasetId, groupId)
}
\arguments{
\item{datasetId}{}

\item{groupId}{The workspace id}
}
\value{
A `data.frame` object.
}
\description{
Updates the datasources of the specified dataset from the specified workspace.Important:
}
